% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIDs.R
\name{AddIDs}
\alias{AddIDs}
\title{AddIDs}
\usage{
AddIDs(method, filtered_genelist, miR_IDs, mRNA_IDs)
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{filtered_genelist}{A list of nested dataframes if 'c' or A list of
lists with nested dataframes if 's'.}

\item{miR_IDs}{miR_ensembl or miR_entrez. Use getIDs function to acquire
this.}

\item{mRNA_IDs}{mRNA_ensembl or mRNA_entrez. Use getIDs function to acquire
this.}
}
\value{
list of dataframes with entrezIDs and genenames additional
as columns which can be stored in metadata section of an MAE.
}
\description{
Adds entrez or ensembl IDs to the nested dataframes within
the filtered_genelist.
}
\examples{
library(clusterProfiler)
library(org.Mm.eg.db)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE
getIDs_miR_mouse(MAE = MAE, miR = MAE@ExperimentList$miR) -> MAE
getIDs_mRNA_mouse(MAE = MAE, mRNA = MAE@ExperimentList$mRNA) -> MAE

CombineGenes(miR_data = MAE@ExperimentList$miR, 
mRNA_data = MAE@ExperimentList$mRNA) -> MAE@ExperimentList$genetic_data

GenesList(method = 'c', genetic_data = MAE@ExperimentList$genetic_data,
timeString = 'D') -> MAE@metadata$genelist

SignificantVals(method = "c", geneList = MAE@metadata$genelist, 
maxVal = 0.05, stringVal = "adjPVal") -> MAE@metadata$filtered_genelist

AddIDs(method = "c", filtered_genelist = MAE@metadata$filtered_genelist, 
miR_IDs = MAE@ExperimentList$miR_entrez,
mRNA_IDs = MAE@ExperimentList$mRNA_entrez) -> MAE@metadata$gene_entrez
}
