\name{c_matrix}
\alias{c_matrix}
\docType{data}
\title{
c_matrix
}
\description{
Matrix of cluster information for the vignette.
}
\usage{data("c_matrix")}
\format{
\describe{
\item{\code{1}}{a numeric vector}
\item{\code{2}}{a numeric vector}
\item{\code{3}}{a numeric vector}
\item{\code{4}}{a numeric vector}
\item{\code{5}}{a numeric vector}
\item{\code{6}}{a numeric vector}
\item{\code{7}}{a numeric vector}
\item{\code{8}}{a numeric vector}
\item{\code{9}}{a numeric vector}
\item{\code{10}}{a numeric vector}
}
}
\details{
Matrix of cluster to wikipathways information.
}
\source{
From data.
}
\references{
NA
}
\examples{
data(c_matrix)
}
\keyword{datasets}
