% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGenes.R
\name{combineGenes}
\alias{combineGenes}
\title{combineGenes}
\usage{
combineGenes(MAE, miR_data, mRNA_data)
}
\arguments{
\item{MAE}{Input MAE object will have the results of combineGenes added to
it as an assay. It is recommended to use the MAE object which stores output
from startObject.}

\item{miR_data}{microRNA dataframe stored as an assay in a MAE. Rows
should be genes, columns are DE results and time point. This should be
the stored within the MAE used in the startObject function.}

\item{mRNA_data}{mRNA dataframe stored as an assay in a MAE. Rows
should be genes, columns are DE results and time point. This should be
the stored within the MAE used in the startObject function.}
}
\value{
A dataframe which combines miR and mRNA data.
}
\description{
Combines miR and mRNA data into one dataframe. Input columns
should be written as :timepoint.DifferentialExpressionResultType e.g.
D1.log2fc or H6.adjPval. Column names should be the same for miR and mRNA
data. A more detailed explanation of column nomenclature can be found in the
vignette. Resulting information will be stored as an assay within a MAE
object. combineGenes is essential for combined analysis of miR-mRNA data.
If using separate analysis, there is no need to use combineGenes.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

mRNA <- mm_mRNA

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- combineGenes(MAE = MAE, miR_data = assay(MAE, 1),
                    mRNA_data = assay(MAE, 2))
}
