% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrTable.R
\name{corrTable}
\alias{corrTable}
\title{corrTable}
\usage{
corrTable(names, miR_exprs, mRNA_wiki, maxInt)
}
\arguments{
\item{names}{Dataframe with combined list of mRNA(single wikipathway of
interest) and miR names.}

\item{miR_exprs}{miR averaged expression data.}

\item{mRNA_wiki}{Single wikipathway of interest's mRNA averaged expression
data.}

\item{maxInt}{Maximum number of data samples in your datasets e.g. how many
timepoints.}
}
\value{
A matrix of each mRNA and miR interaction and a correlation.
}
\description{
corrTable
}
