% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoMake.R
\name{CytoMake}
\alias{CytoMake}
\title{CytoMake}
\usage{
CytoMake(interaction_data, titleString = '', collectionString = '')
}
\arguments{
\item{interaction_data}{Output from MatrixFiler. Requires a dataframe with
columns names MIR and mRNA}

\item{titleString}{Title of the network.}

\item{collectionString}{Title of the collection fo networks.}
}
\value{
A network visible in cytoscape.
}
\description{
Creates a cytoscape network based on the output of MatrixFilter.
Requires RCy3 package and cytoscapePing() to be used. Make sure Cytoscape
is open first.
}
