% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMiningMatrix.R
\name{DataMiningMatrix}
\alias{DataMiningMatrix}
\title{DataMiningMatrix}
\usage{
DataMiningMatrix(interactionsDF, targetscanInt ,
mirdbInt, mirtarbaseInt)
}
\arguments{
\item{interactionsDF}{Correlation matrix between the mRNAs from the
wikipathway of interest and miRNA data.}

\item{targetscanInt}{Targetscans predicted interactions.}

\item{mirdbInt}{miRDB predcited interactions.}

\item{mirtarbaseInt}{miRTarBase functional interactions.}
}
\value{
A matrix of information which cross references the occurance of
microRNA-mRNA interactions between databases and the given data.
}
\description{
Mines out predicted/ functional interactions which correspond
between Targetscans, miRDB, miRTarBase and the interactions data frame.
}
