% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadGmt.R
\name{dloadGmt}
\alias{dloadGmt}
\title{dloadGmt}
\usage{
dloadGmt(MAE, species = "")
}
\arguments{
\item{MAE}{MultiAssayExperiment to store downloaded GMT data in. It might
be useful to start a new MAE for dloadGmt using MultiAssayExperiment(). This
is so the MAE objects used in this analysis do not get too large.}

\item{species}{Full species names e.g. retrieve "Homo sapiens" or "Mus musculus" data.}
}
\value{
3 dataframes. 1) path_gene, 2) path_names, 3) path_data.
All of which will be stored as assays in the input MAE.
}
\description{
Downloads the most up-to-date versions of the mouse or human
wikipathway information databases. Output will be stored as three distinct
dataframes within the input MAE 1) path_gene, 2) path_names, 3) path_data.
}
\examples{
MAE <- MultiAssayExperiment()

MAE <- dloadGmt(MAE, species = "Homo sapiens")
}
