% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadMirdb.R
\name{dloadMirdb}
\alias{dloadMirdb}
\title{dloadMirdb}
\usage{
dloadMirdb(MAE, species, orgDB)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store downloaded mirDB data.
It is recommended to use the MAE which was used in the mirMrnaInt function.}

\item{species}{Species of interest e.g. "hsa" or "mmu".}

\item{orgDB}{Organism db package specific for species of interest.}
}
\value{
A dataframe of predicted, species specific mRNA-miR interactions.
Will be stored as an assay in the input MAE.
}
\description{
Downloads most recent version (6.0) of predicted targets from
the mirdb database http://mirdb.org/download.html. This will take some time.
miR-mRNA interactions from the species of interest will be extracted.
Species of interests associated org.db package must be loaded beforehand.
}
\examples{
\dontrun{

library(org.Mm.eg.db)

MAE <- MultiAssayExperiment()

MAE <-dloadMirdb(MAE, 'mmu', org.Mm.eg.db)

}
}
