% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadTargetscan.R
\name{dloadTargetscan}
\alias{dloadTargetscan}
\title{dloadTargetscan}
\usage{
dloadTargetscan(MAE, species)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will have the downloaded targetscans
data stored here. It is recommended to use the same MAE object created from
the mirMrnaInt function.}

\item{species}{Species of interest e.g "hsa" or "mmu."}
}
\value{
Dataframe of species specific predicted mRNA-miR interactions.

A dataframe of predicted miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (7.2) of predicted targets from
targetscan database
http://www.targetscan.org/cgi-bin/targetscan/data_download.vert72.cgi.
Resulting dataframe will be stored as an assay in a MAE object.
}
\examples{
\dontrun{

MAE <- MultiAssayExperiment()

MAE <-dloadTargetscan(MAE, "mmu")
}
}
