% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadTargetScans.R
\name{dloadTargetScans}
\alias{dloadTargetScans}
\title{dloadTargetScans}
\usage{
dloadTargetScans()
}
\value{
A dataframe of predicted miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (7.2) of predicted targets from
targetscan database.
}
