% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGMT.R
\name{downloadGMT}
\alias{downloadGMT}
\title{downloadGMT}
\usage{
downloadGMT(speciesInitials = "")
}
\arguments{
\item{speciesInitials}{Either Hs or Mm, respectively to retreive Homo
sapiens or Mus musculus data.}
}
\value{
3 dataframes. 1) genes-wikipathwayIDs 2) wikipathwayIDs-wikipathway
fullnames 3) combination of 1 and 2.
}
\description{
Provides mouse or human wikipathway information.
}
\examples{
downloadGMT(speciesInitial = "Mm")
file.remove("mus.gmt")
}
