\name{e_list}
\alias{e_list}
\docType{data}
\title{
e_list
}
\description{
List of entrezIDs to make examples run faster.
}
\usage{data("e_list")}
\format{
The format is:
List of 5
$ D1 : Factor w/ 2157 levels "100049549","100124432",..:
1007 1820 1294 1458 1385 294 295 1062 1295 241 ...
$ D2 : Factor w/ 2157 levels "100049549","100124432",..:
1225 1820 1294 1458 1385 294 295 1062 2063 1295 ...
$ D3 : Factor w/ 2157 levels "100049549","100124432",..:
1225 1294 1458 1385 294 2063 1295 241 210 296 ...
$ D7 : Factor w/ 2157 levels "100049549","100124432",..:
1458 1385 294 425 2063 874 1365 241 1771 296 ...
$ D14: Factor w/ 2157 levels "100049549","100124432",..:
241 1379 274 308 1040 498 533 879 321 1930 ...
}
\details{
List of entrezgeneIDs per time point in data.
}
\source{
From data.
}
\references{
NA
}
\examples{
data(e_list)
}
\keyword{datasets}
