% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNames.R
\name{eNames}
\alias{eNames}
\title{eNames}
\usage{
eNames(method = '', gene_IDs, ID_Column)
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{gene_IDs}{A nested list of dataframes(c)/ list of lists of nested
dataframes (s), and each dataframe has a column for entrez or ensembl ID.}

\item{ID_Column}{The entrez/ ensembl ID column in each dataframe.
Should be the last column.}
}
\value{
A single list of entrez/ ensembl IDs for each dataframe
within the lists.
}
\description{
Extract the ID names from the nested dataframes.
}
\examples{
library(biomaRt)
library(clusterProfiler)
library(org.Mm.eg.db)
miR <- mm_miR
miR <- miR[1:100,]
mRNA <- mm_mRNA
mRNA <- mRNA[1:200,]
getIDs_miR_mouse(miR = miR)
getIDs_mRNA_mouse(mRNA = mRNA)
CombineGenes(miR_data = miR, mRNA_data = mRNA) -> genetic_data
GenesList(method = 'c', genetic_data = genetic_data,
timeString = 'D') -> genelist
SignificantVals(method = 'c', geneList = genelist, maxVal = 0.05,
stringVal = 'adjPVal') -> filtered_genelist
AddIDs(method = 'c', filtered_genelist = filtered_genelist,
miR_IDs = miR_entrez, mRNA_IDs = mRNA_entrez) -> gene_entrez
eNames(method = 'c', gene_IDs = gene_entrez, ID_Column = 4) -> e_list
}
