% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichWiki.R
\name{EnrichWiki}
\alias{EnrichWiki}
\title{EnrichWiki}
\usage{
EnrichWiki(method = '', e_list, orgDB, path_gene, path_name, ID = '',
universe, pvalcutoff, qvaluecutoff, padjustmethod)
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{e_list}{List of ensembl or entrez IDs for each sample.}

\item{orgDB}{Library of species specific data.}

\item{path_gene}{genes-wikipathwayIDs dataframe. From downladGMT or
GMT_ensembl.}

\item{path_name}{wikipathwayIDs-wikipathway fullnames dataframe.
From downladGMT.}

\item{ID}{Either ENTREZID or ENSEMBL.}

\item{universe}{A column of gene IDs to be used as the background for
gene set enrichment. To find which pathways are most enriched use
path_gene$gene}

\item{pvalcutoff}{Default is 0.05. P value cutuff point.}

\item{qvaluecutoff}{Default is 0.2. q value cutoff point.}

\item{padjustmethod}{Default is 'BH'. Which pvalue adjustment method should
be used. Look into clusterProfiler function enricher for more options.}
}
\value{
A large list of data which identifies which wikipathways are
most enriched in the input data.
}
\description{
Finds which wikipathways are enriched within the data.
}
\examples{
library(org.Mm.eg.db)
library(clusterProfiler)
e_list -> elist
downloadGMT(speciesInitial = "Mm")
file.remove("mus.gmt")
EnrichWiki(method = 'c', elist, org.Mm.eg.db,
path_gene, path_name, ID = 'ENTREZID',
universe = path_gene$gene) -> sigwiki
}
