% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Express.R
\name{Express}
\alias{Express}
\title{Express}
\usage{
Express(df, dataType = '', genes_ID, idColumn = '')
}
\arguments{
\item{df}{mRNA or miR dataframe (rownames as genes and experimental
variables as columns.)}

\item{dataType}{Column name to take an average from e.g. Log2FC, AveExp,
pval.}

\item{genes_ID}{Dataframe with that was created earlier e.g. mRNA_ensembl
or miR_entrez.}

\item{idColumn}{Name of column to use as the merge point.}
}
\value{
Returns the average of whichever data the user was interested e.g.
Average expression or average log2fc.
}
\description{
Get average data for one result type.
}
\examples{
library(biomaRt)
mm_mRNA -> mRNA
getIDs_mRNA_mouse(mRNA)
Express(df = mRNA, dataType = 'Log2FC', genes_ID = mRNA_entrez,
idColumn = 'GENENAME') -> mRNA_express
}
