% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenesList.R
\name{GenesList}
\alias{GenesList}
\title{GenesLists}
\usage{
GenesList(method, genetic_data, timeString, miR_data, mRNA_data)
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{genetic_data}{If 'c', this should be a dataframe with miR and mRNA
information together.}

\item{timeString}{If 'c', this should be a commmon string representing
'timepoint' e.g. for H.1, H.10, H.20, timeString = 'H'.}

\item{miR_data}{If 's', a dataframe of microRNA data. Rownames are genes
and colnames are like so; genetype_timepoint.resulttype.
timepoint.resulttype should be the same in mRNA and miR data.}

\item{mRNA_data}{If 's', a dataframe of mRNA data. Rownames are genes
and colnames are like so; genetype_timepoint.resulttype.
timepoint.resulttype should be the same in mRNA and miR data.}
}
\value{
A list of dataframes separated by features in the column names which
can be stored in the metadata area of an MAE object.
}
\description{
Will produce a list of nested dataframes. For combined
analysis method = c, and for separated analysis method = s. Combined
analysis colnames should be as follows 'timepoint.resultstype'.
GenesList will make new dataframes separated by 'timepoint.'.
Separated colnames should be as follows 'genetype_timepoint.resulttype'.
GenesList will make separate lists for 'genetype_'. Each of these lists
will have dataframes which have been made by separating by 'timepoint.'.
Make sure to follow colname nomenclature carefully.
}
\examples{
mm_miR -> miR
mm_mRNA -> mRNA

StartObject(miR = miR, mRNA = mRNA) -> MAE

AddPrefix(gene_df = MAE@ExperimentList$miR,
prefixString = "miR" ) -> MAE@ExperimentList$miR_p
AddPrefix(gene_df = MAE@ExperimentList$mRNA, 
prefixString = "mRNA") -> MAE@ExperimentList$mRNA_p

GenesList(method = "s", miR_data = MAE@ExperimentList$miR_p,
mRNA_data = MAE@ExperimentList$mRNA_p) -> MAE@metadata$genelist

CombineGenes(miR_data = MAE@ExperimentList$miR, mRNA_data = 
MAE@ExperimentList$mRNA) -> MAE@ExperimentList$genetic_data

GenesList(method = 'c', genetic_data = MAE@ExperimentList$genetic_data,
timeString = 'D') -> MAE@metadata$genelist
}
