% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_miR_mouse.R
\name{getIDs_miR_mouse}
\alias{getIDs_miR_mouse}
\title{getIDs_miR_mouse}
\usage{
getIDs_miR_mouse(miR)
}
\arguments{
\item{miR}{Dataframe. Rownames are genes and columns are results of DE.}
}
\value{
4 dataframes consisting of ID information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data for mouse
microRNAs. It will aslo produce adjusted ensembl and entrez for IDs that
are specific to multiple microRNAs.
}
\examples{
library(clusterProfiler)
library(org.Mm.eg.db)
mm_miR -> miR
getIDs_miR_mouse(miR)
}
