% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_miR_mousetohuman.R
\name{getIDs_miR_mousetohuman}
\alias{getIDs_miR_mousetohuman}
\title{getIDs_miR_mousetohuman}
\usage{
getIDs_miR_mousetohuman(miR, mirror = "useast")
}
\arguments{
\item{miR}{Dataframe. Rownames should be genenames and colnames should be
results from DE.}

\item{mirror}{String to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
5 dataframes consisting of ID information, and a new dataframe of
human named miRs with the DE results.
}
\description{
getIDs_miR_mousetohuman will produce ensembl and entrez data
for mouse microRNAs. It will aslo produce adjusted ensembl and entrez for
IDs that are specific to multiple microRNAs. Finally a new miR_human_named
file will be made.
}
\examples{
library(biomaRt)
library(clusterProfiler)
library(org.Hs.eg.db)
mm_miR -> miR
miR[1:100,] -> miR
getIDs_miR_mousetohuman(miR = miR, mirror = 'useast')
}
