% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_mRNA_human.R
\name{getIDs_mRNA_human}
\alias{getIDs_mRNA_human}
\title{getIDs_mRNA_human}
\usage{
getIDs_mRNA_human(MAE, mRNA, mirror)
}
\arguments{
\item{MAE}{MultiAssayObject created by StartObject}

\item{mRNA}{Dataframe. Rownames should be genenames and colnames should
be results from DE.}

\item{mirror}{String to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
2 new dataframes. One with entrez information and another with
ensembl information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data
for human mRNAs.
}
\examples{
library(biomaRt)
library(MultiAssayExperiment)
mRNA <- mm_mRNA
mRNA <- mRNA[1:20,]
MAE <- StartObject(miR = NULL, mRNA = mRNA)
MAE <- getIDs_mRNA_human(MAE = MAE, mRNA = assay(MAE, 2), mirror = 'useast')
}
