% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_mRNA_mouse.R
\name{getIDs_mRNA_mouse}
\alias{getIDs_mRNA_mouse}
\title{getIDs_mRNA_mouse}
\usage{
getIDs_mRNA_mouse(mRNA, mirror)
}
\arguments{
\item{mRNA}{Dataframe. Rownames should be genenames and colnames should be
results from DE.}

\item{mirror}{tring to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
2 new dataframes. One with entrez information and another with
ensembl information.
}
\description{
getIDs_miR_mouse will produce ensembl and entrez data for
mouse mRNAs.
}
\examples{
library(biomaRt)
mm_mRNA -> mRNA
mRNA[1:200,] -> mRNA
getIDs_mRNA_mouse(mRNA)
}
