% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMirHuman.R
\name{getIdsMirHuman}
\alias{getIdsMirHuman}
\title{getIdsMirHuman}
\usage{
getIdsMirHuman(MAE, miR)
}
\arguments{
\item{MAE}{MultiAssayExperiment to store the output of getIdsMirHuman.
It is recommended to use the MAE which contains output from startObject.}

\item{miR}{Dataframe. Rownames are genes and columns are results of DE.
This should be found as an assay within the MAE used in the
startObject function. Please read vignette for nomenclature guidance.}
}
\value{
MAE object with 4 more dataframes consisting of ID information.
}
\description{
getIdsMirHuman will produce ensembl and entrez ID data for
human microRNAs. It will also produce adjusted ensembl and entrez for IDs
that are specific to microRNAs that share an ID. They will be stored as
4 individual assays in a MAE. org.Hs.eg.db must be loaded prior to use of
this function.
}
\examples{
library(org.Hs.eg.db)

miR <- hs_miR

# Make sure miRNA gene name nomenclature is correct for TimiRGeN analysis!

rownames(miR) <- gsub(rownames(miR), pattern = "\\\\.", replacement = "-")

rownames(miR) <- sub("-$", "*", rownames(miR))

miR <- miR[1:100,]

MAE <- startObject(miR = miR, mRNA = NULL)

MAE <- getIdsMirHuman(MAE, assay(MAE, 1))
}
