% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMirMouse.R
\name{getIdsMirMouse}
\alias{getIdsMirMouse}
\title{getIdsMirMouse}
\usage{
getIdsMirMouse(MAE, miR)
}
\arguments{
\item{MAE}{MultiAssayObject created by StartObject}

\item{miR}{Dataframe. Rownames are genes and columns are results of DE.}
}
\value{
MAE object with 4 more dataframes consisting of ID information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data for mouse
microRNAs. It will aslo produce adjusted ensembl and entrez for IDs that
are specific to microRNAs that share an ID.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

miR <- miR[1:100,]

MAE <- startObject(miR = miR, mRNA = NULL)

MAE <- getIdsMirMouse(MAE, assay(MAE, 1))
}
