% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMrna.R
\name{getIdsMrna}
\alias{getIdsMrna}
\title{getIdsMrna}
\usage{
getIdsMrna(MAE, mRNA, mirror, species)
}
\arguments{
\item{MAE}{MultiAssayExperiment to store the output of getIdsMrna.
It is recommended to use the MAE which contains output from startObject.}

\item{mRNA}{Dataframe. Rownames are genes and columns are results of DE.
This should be found as an assay within the MAE used in the startObject
function. Please read vignette for nomenclature guidance.}

\item{mirror}{String to identify which biomaRt server is best. This is based
on location. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'www'.}

\item{species}{Species of interest which includes}
}
\value{
2 new dataframes in the MAE. One with entrez information and
the other with ensembl gene ID information.
}
\description{
getIdsMrna will produce ensembl and entrez ID dataframes
for mRNAs. These will be stored as 2 individual assays within a MAE.
}
\examples{
mRNA <- mm_mRNA

mRNA <- mRNA[1:20,]

MAE <- startObject(miR = NULL, mRNA = mRNA)

MAE <- getIdsMrna(MAE = MAE, mRNA = assay(MAE, 2), mirror = 'useast',
                      species = 'mmusculus')
}
