% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMrnaHuman.R
\name{getIdsMrnaHuman}
\alias{getIdsMrnaHuman}
\title{getIdsMrnaHuman}
\usage{
getIdsMrnaHuman(MAE, mRNA, mirror)
}
\arguments{
\item{MAE}{MultiAssayObject created by StartObject}

\item{mRNA}{Dataframe. Rownames should be genenames and colnames should
be results from DE and time points.}

\item{mirror}{String to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
2 new dataframes. One with entrez information and another with
ensembl information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data
for human mRNAs.
}
\examples{
mRNA <- mm_mRNA

mRNA <- mRNA[1:20,]

MAE <- startObject(miR = NULL, mRNA = mRNA)

MAE <- getIdsMrnaHuman(MAE = MAE, mRNA = assay(MAE, 2), mirror = 'useast')
}
