% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMrnaMouse.R
\name{getIdsMrnaMouse}
\alias{getIdsMrnaMouse}
\title{getIdsMrnaMouse}
\usage{
getIdsMrnaMouse(MAE, mRNA, mirror)
}
\arguments{
\item{MAE}{MultiAssayObject created by StartObject}

\item{mRNA}{Dataframe. Rownames should be genenames and colnames should be
results from DE and time points.}

\item{mirror}{tring to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
2 new dataframes in the MAE object. One with entrez information and
the other with ensembl information.
}
\description{
getIDs_miR_mouse will produce ensembl and entrez data for
mouse mRNAs.
}
\examples{
mRNA <- mm_mRNA

mRNA <- mRNA[1:20,]

MAE <- startObject(miR = NULL, mRNA = mRNA)

MAE <- getIdsMrnaMouse(MAE = MAE, mRNA = assay(MAE, 2), mirror = 'useast')
}
