% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT_ensembl.R
\name{GMT_ensembl}
\alias{GMT_ensembl}
\title{GMT_ensembl}
\usage{
GMT_ensembl(MAE, path_gene, path_data, orgDB)
}
\arguments{
\item{MAE}{MultiAssayExperiment where the ensembl wikipathways data can 
be stored.}

\item{path_gene}{wikipathway-gene data from downloadGMT.}

\item{path_data}{wikipathway-gene-wikipathway full name data from
downloadGMT.}

\item{orgDB}{organism Library e.g. org.Hs.eg.db}
}
\value{
2 dataframes. Wikipathway data based on ensembl IDs.
}
\description{
Change entrez IDs in wpid2gene and wp_data into ensembl IDs
}
\examples{
library(MultiAssayExperiment)
library(org.Mm.eg.db)
miR <- mm_miR
mRNA <- mm_mRNA
MAE <- StartObject(miR = miR, mRNA = mRNA)
MAE <- dloadGMT(MAE, speciesInitial = "Mm")

MAE <- GMT_ensembl(MAE = MAE, assay(MAE, 3),
                   assay(MAE, 5), org.Mm.eg.db)
}
