\name{Int_data}
\alias{Int_data}
\docType{data}
\title{
Int_data
}
\description{
Data frame of the correlations found between the DE lof2fc scores from the
miR and mRNA data from Mouse Kidney Fibrosis data set. Includes if these
interactions were also found in TargetScans, miRDB and miRTarBase.
}
\usage{data("Int_data")}
\format{
A data frame with 6394 observations on the following 10 variables.
\describe{
\item{\code{avecor}}{a numeric vector}
\item{\code{miR}}{a factor with levels \code{mmu-let-7b-3p} \code{mmu-let-7c-5p} \code{mmu-let-7d-3p} \code{mmu-let-7e-3p} \code{mmu-let-7e-5p} \code{mmu-let-7f-1-3p} \code{mmu-let-7g-5p} \code{mmu-let-7i-3p} \code{mmu-let-7i-5p} \code{mmu-let-7j} \code{mmu-miR-101a-3p} \code{mmu-miR-101a-5p} \code{mmu-miR-101c} \code{mmu-miR-106a-5p} \code{mmu-miR-106b-3p} \code{mmu-miR-106b-5p} \code{mmu-miR-107-3p} \code{mmu-miR-10a-3p} \code{mmu-miR-10a-5p} \code{mmu-miR-1191a} \code{mmu-miR-1199-5p} \code{mmu-miR-1249-3p} \code{mmu-miR-125a-3p} \code{mmu-miR-125b-1-3p} \code{mmu-miR-125b-2-3p} \code{mmu-miR-126a-3p} \code{mmu-miR-126a-5p} \code{mmu-miR-127-3p} \code{mmu-miR-128-3p} \code{mmu-miR-130b-3p} \code{mmu-miR-130b-5p} \code{mmu-miR-132-3p} \code{mmu-miR-132-5p} \code{mmu-miR-133a-3p} \code{mmu-miR-138-5p} \code{mmu-miR-139-5p} \code{mmu-miR-141-3p} \code{mmu-miR-141-5p} \code{mmu-miR-142a-3p} \code{mmu-miR-142a-5p} \code{mmu-miR-144-5p} \code{mmu-miR-145a-3p} \code{mmu-miR-145a-5p} \code{mmu-miR-146a-5p} \code{mmu-miR-146b-5p} \code{mmu-miR-147-3p} \code{mmu-miR-148a-3p} \code{mmu-miR-148a-5p} \code{mmu-miR-149-5p} \code{mmu-miR-151-5p} \code{mmu-miR-152-5p} \code{mmu-miR-155-5p} \code{mmu-miR-15a-5p} \code{mmu-miR-15b-3p} \code{mmu-miR-15b-5p} \code{mmu-miR-16-1-3p} \code{mmu-miR-16-5p} \code{mmu-miR-17-3p} \code{mmu-miR-17-5p} \code{mmu-miR-181a-1-3p} \code{mmu-miR-181a-2-3p} \code{mmu-miR-181a-5p} \code{mmu-miR-181b-5p} \code{mmu-miR-181c-3p} \code{mmu-miR-181c-5p} \code{mmu-miR-181d-5p} \code{mmu-miR-182-5p} \code{mmu-miR-183-5p} \code{mmu-miR-1843a-5p} \code{mmu-miR-1843b-5p} \code{mmu-miR-184-3p} \code{mmu-miR-185-5p} \code{mmu-miR-187-3p} \code{mmu-miR-187-5p} \code{mmu-miR-188-5p} \code{mmu-miR-1895} \code{mmu-miR-18a-3p} \code{mmu-miR-18a-5p} \code{mmu-miR-190a-3p} \code{mmu-miR-190a-5p} \code{mmu-miR-190b-5p} \code{mmu-miR-191-5p} \code{mmu-miR-192-5p} \code{mmu-miR-193a-3p} \code{mmu-miR-193a-5p} \code{mmu-miR-193b-3p} \code{mmu-miR-194-5p} \code{mmu-miR-1948-3p} \code{mmu-miR-195a-5p} \code{mmu-miR-1968-5p} \code{mmu-miR-196a-1-3p} \code{mmu-miR-196a-2-3p} \code{mmu-miR-196b-5p} \code{mmu-miR-1981-3p} \code{mmu-miR-199a-3p} \code{mmu-miR-199a-5p} \code{mmu-miR-199b-3p} \code{mmu-miR-199b-5p} \code{mmu-miR-19a-3p} \code{mmu-miR-19b-3p} \code{mmu-miR-200a-3p} \code{mmu-miR-200a-5p} \code{mmu-miR-200c-3p} \code{mmu-miR-203-3p} \code{mmu-miR-203b-3p} \code{mmu-miR-204-5p} \code{mmu-miR-20a-5p} \code{mmu-miR-20b-5p} \code{mmu-miR-210-5p} \code{mmu-miR-212-3p} \code{mmu-miR-212-5p} \code{mmu-miR-2137} \code{mmu-miR-214-3p} \code{mmu-miR-214-5p} \code{mmu-miR-21a-3p} \code{mmu-miR-21a-5p} \code{mmu-miR-221-3p} \code{mmu-miR-223-3p} \code{mmu-miR-223-5p} \code{mmu-miR-22-3p} \code{mmu-miR-224-5p} \code{mmu-miR-22-5p} \code{mmu-miR-23a-3p} \code{mmu-miR-23b-3p} \code{mmu-miR-24-2-5p} \code{mmu-miR-24-3p} \code{mmu-miR-25-3p} \code{mmu-miR-25-5p} \code{mmu-miR-26a-2-3p} \code{mmu-miR-26a-5p} \code{mmu-miR-26b-3p} \code{mmu-miR-26b-5p} \code{mmu-miR-27a-3p} \code{mmu-miR-27a-5p} \code{mmu-miR-27b-3p} \code{mmu-miR-28a-3p} \code{mmu-miR-298-5p} \code{mmu-miR-29a-3p} \code{mmu-miR-29b-2-5p} \code{mmu-miR-29c-3p} \code{mmu-miR-300-3p} \code{mmu-miR-301a-3p} \code{mmu-miR-301b-3p} \code{mmu-miR-3057-5p} \code{mmu-miR-3061-3p} \code{mmu-miR-3073b-3p} \code{mmu-miR-3073b-5p} \code{mmu-miR-30a-3p} \code{mmu-miR-30a-5p} \code{mmu-miR-30b-5p} \code{mmu-miR-30c-1-3p} \code{mmu-miR-30c-2-3p} \code{mmu-miR-30c-5p} \code{mmu-miR-30d-3p} \code{mmu-miR-30d-5p} \code{mmu-miR-30e-3p} \code{mmu-miR-30e-5p} \code{mmu-miR-3105-3p} \code{mmu-miR-31-3p} \code{mmu-miR-31-5p} \code{mmu-miR-320-3p} \code{mmu-miR-322-3p} \code{mmu-miR-324-5p} \code{mmu-miR-326-3p} \code{mmu-miR-328-3p} \code{mmu-miR-330-3p} \code{mmu-miR-330-5p} \code{mmu-miR-331-3p} \code{mmu-miR-33-3p} \code{mmu-miR-339-3p} \code{mmu-miR-339-5p} \code{mmu-miR-342-3p} \code{mmu-miR-342-5p} \code{mmu-miR-344-3p} \code{mmu-miR-345-3p} \code{mmu-miR-34a-5p} \code{mmu-miR-34b-3p} \code{mmu-miR-34b-5p} \code{mmu-miR-34c-3p} \code{mmu-miR-34c-5p} \code{mmu-miR-350-3p} \code{mmu-miR-350-5p} \code{mmu-miR-351-5p} \code{mmu-miR-3535} \code{mmu-miR-361-5p} \code{mmu-miR-362-3p} \code{mmu-miR-363-3p} \code{mmu-miR-365-2-5p} \code{mmu-miR-365-3p} \code{mmu-miR-374b-5p} \code{mmu-miR-374c-5p} \code{mmu-miR-375-3p} \code{mmu-miR-378a-3p} \code{mmu-miR-378a-5p} \code{mmu-miR-378b} \code{mmu-miR-378c} \code{mmu-miR-378d} \code{mmu-miR-3964} \code{mmu-miR-3965} \code{mmu-miR-3966} \code{mmu-miR-3968} \code{mmu-miR-3969} \code{mmu-miR-409-3p} \code{mmu-miR-423-3p} \code{mmu-miR-423-5p} \code{mmu-miR-425-3p} \code{mmu-miR-429-3p} \code{mmu-miR-450a-5p} \code{mmu-miR-451a} \code{mmu-miR-455-3p} \code{mmu-miR-455-5p} \code{mmu-miR-466a-3p} \code{mmu-miR-466b-3p} \code{mmu-miR-466c-3p} \code{mmu-miR-466e-3p} \code{mmu-miR-466p-3p} \code{mmu-miR-467a-5p} \code{mmu-miR-467b-5p} \code{mmu-miR-467c-5p} \code{mmu-miR-467e-5p} \code{mmu-miR-470-5p} \code{mmu-miR-484} \code{mmu-miR-486a-3p} \code{mmu-miR-486a-5p} \code{mmu-miR-486b-3p} \code{mmu-miR-486b-5p} \code{mmu-miR-497a-5p} \code{mmu-miR-500-3p} \code{mmu-miR-5099} \code{mmu-miR-511-3p} \code{mmu-miR-5129-3p} \code{mmu-miR-532-3p} \code{mmu-miR-532-5p} \code{mmu-miR-541-5p} \code{mmu-miR-574-3p} \code{mmu-miR-574-5p} \code{mmu-miR-582-3p} \code{mmu-miR-598-3p} \code{mmu-miR-615-3p} \code{mmu-miR-615-5p} \code{mmu-miR-6539} \code{mmu-miR-664-3p} \code{mmu-miR-669a-3p} \code{mmu-miR-669a-5p} \code{mmu-miR-669c-5p} \code{mmu-miR-669o-3p} \code{mmu-miR-669p-5p} \code{mmu-miR-671-3p} \code{mmu-miR-671-5p} \code{mmu-miR-672-5p} \code{mmu-miR-674-3p} \code{mmu-miR-674-5p} \code{mmu-miR-676-3p} \code{mmu-miR-676-5p} \code{mmu-miR-677-5p} \code{mmu-miR-690} \code{mmu-miR-6944-3p} \code{mmu-miR-6978-3p} \code{mmu-miR-700-5p} \code{mmu-miR-744-3p} \code{mmu-miR-7a-1-3p} \code{mmu-miR-802-3p} \code{mmu-miR-802-5p} \code{mmu-miR-8112} \code{mmu-miR-8114} \code{mmu-miR-874-3p} \code{mmu-miR-874-5p} \code{mmu-miR-877-5p} \code{mmu-miR-92a-1-5p} \code{mmu-miR-92a-3p} \code{mmu-miR-92b-3p} \code{mmu-miR-93-3p} \code{mmu-miR-93-5p} \code{mmu-miR-96-5p} \code{mmu-miR-98-5p} \code{mmu-miR-99a-3p} \code{mmu-miR-99a-5p} \code{mmu-miR-99b-3p}}
\item{\code{mRNA}}{a factor with levels \code{A2m} \code{C1qa} \code{C1qb} \code{C1qc} \code{C1ra} \code{C3} \code{C3ar1} \code{C4b} \code{C6} \code{C7} \code{C8g} \code{Cd55} \code{Cfh} \code{Cfi} \code{F13b} \code{F2r} \code{F3} \code{F5} \code{Masp1} \code{Plaur} \code{Proc} \code{Serpine1} \code{Serping1}}
\item{\code{miR_Entrez}}{a factor with levels \code{100049549} \code{100124432} \code{100124448} \code{100124451} \code{100124452} \code{100124459} \code{100124470} \code{100124481} \code{100124487} \code{100124488} \code{100124489} \code{100124491} \code{100124640} \code{100316677} \code{100316699} \code{100316702} \code{100316713} \code{100316737} \code{100316779} \code{100316821} \code{100316832} \code{100526461} \code{100526465} \code{100526467} \code{100526497} \code{100526513} \code{100526530} \code{100526562} \code{100628570} \code{100628571} \code{100628572} \code{100628573} \code{100628578} \code{100628598} \code{100628607} \code{100628616} \code{100628622} \code{100628623} \code{100885836} \code{102465206} \code{102465224} \code{102465310} \code{102465569} \code{102465591} \code{102465900} \code{102466257} \code{387134} \code{387139} \code{387140} \code{387141} \code{387143} \code{387145} \code{387146} \code{387150} \code{387157} \code{387159} \code{387162} \code{387163} \code{387165} \code{387166} \code{387167} \code{387169} \code{387170} \code{387173} \code{387174} \code{387175} \code{387176} \code{387177} \code{387178} \code{387179} \code{387180} \code{387182} \code{387183} \code{387185} \code{387186} \code{387187} \code{387188} \code{387190} \code{387191} \code{387199} \code{387200} \code{387206} \code{387208} \code{387210} \code{387216} \code{387217} \code{387219} \code{387220} \code{387221} \code{387222} \code{387224} \code{387225} \code{387226} \code{387227} \code{387228} \code{387229} \code{387230} \code{387235} \code{387236} \code{387239} \code{387242} \code{723814} \code{723816} \code{723819} \code{723820} \code{723826} \code{723827} \code{723829} \code{723830} \code{723832} \code{723833} \code{723836} \code{723838} \code{723840} \code{723841} \code{723848} \code{723849} \code{723850} \code{723851} \code{723852} \code{723853} \code{723862} \code{723864} \code{723865} \code{723870} \code{723873} \code{723885} \code{723886} \code{723889} \code{723891} \code{723893} \code{723894} \code{723895} \code{723896} \code{723897} \code{723898} \code{723900} \code{723905} \code{723907} \code{723908} \code{723909} \code{723910} \code{723916} \code{723921} \code{723923} \code{723925} \code{723926} \code{723931} \code{723932} \code{723941} \code{723944} \code{723946} \code{723947} \code{723952} \code{723958} \code{723960} \code{723962} \code{723963} \code{723964} \code{723974} \code{724063} \code{732489} \code{735252} \code{735257} \code{735262} \code{735264} \code{735273} \code{735285} \code{751518} \code{751519} \code{751534} \code{751535} \code{751543} \code{751544} \code{751546} \code{751550} \code{751557} \code{791069} \code{791070} \code{791074}}
\item{\code{mRNA_Entrez}}{a factor with levels \code{100034251} \code{100038538} \code{100038947} \code{100040462} \code{100061} \code{100066} \code{100102} \code{100129} \code{100340} \code{100383} \code{100463512} \code{100504474} \code{100559} \code{100604} \code{100727} \code{100862383} \code{100952} \code{101056186} \code{101488} \code{101497} \code{101613} \code{101631} \code{101809} \code{102103} \code{102294} \code{102448} \code{102570} \code{102632} \code{103172} \code{103573} \code{103742} \code{103743} \code{103768} \code{103967} \code{104080} \code{104130} \code{104215} \code{104445} \code{104718} \code{104759} \code{104776} \code{105014} \code{105243} \code{105245} \code{105348} \code{105349} \code{105355} \code{105387} \code{105445} \code{105785} \code{105855} \code{105887} \code{106064} \code{106572} \code{106585} \code{106639} \code{106766} \code{106861} \code{106877} \code{106931} \code{107146} \code{107515} \code{107566} \code{107581} \code{107766} \code{107771} \code{107869} \code{107971} \code{107995} \code{108000} \code{108012} \code{108037} \code{108075} \code{108098} \code{108114} \code{108151} \code{108672} \code{108912} \code{108961} \code{109082} \code{109154} \code{109225} \code{109294} \code{109342} \code{109620} \code{109624} \code{109648} \code{109652} \code{109711} \code{109828} \code{109901} \code{110006} \code{110033} \code{110175} \code{110198} \code{110213} \code{110310} \code{110351} \code{110391} \code{110446} \code{110695} \code{110826} \code{110842} \code{110877} \code{111175} \code{112417} \code{11287} \code{11303} \code{11304} \code{11364} \code{11370} \code{113868} \code{11416} \code{11421} \code{11438} \code{11443} \code{114584} \code{114644} \code{114664} \code{11484} \code{11488} \code{11492} \code{11501} \code{11504} \code{11513} \code{11520} \code{11522} \code{11529} \code{11568} \code{11606} \code{11628} \code{11629} \code{11668} \code{116847} \code{11690} \code{117147} \code{117160} \code{11727} \code{11740} \code{11745} \code{11796} \code{11797} \code{11810} \code{11816} \code{11829} \code{11835} \code{118449} \code{118454} \code{11852} \code{11854} \code{11855} \code{11858} \code{11865} \code{11867} \code{11872} \code{11881} \code{11898} \code{11906} \code{11910} \code{11936} \code{11946} \code{11947} \code{11949} \code{11950} \code{11951} \code{11957} \code{11988} \code{12010} \code{12021} \code{12036} \code{12039} \code{12040} \code{12048} \code{12051} \code{12053} \code{12111} \code{12153} \code{12169} \code{12189} \code{12192} \code{12224} \code{12235} \code{12236} \code{12258} \code{12259} \code{12260} \code{12262} \code{12265} \code{12266} \code{12267} \code{12268} \code{12274} \code{12293} \code{12297} \code{12298} \code{12306} \code{12316} \code{12321} \code{12332} \code{12351} \code{12359} \code{12362} \code{12374} \code{12386} \code{12394} \code{12424} \code{12425} \code{12442} \code{12484} \code{12489} \code{12491} \code{12496} \code{12499} \code{12505} \code{12508} \code{12534} \code{12544} \code{12552} \code{12563} \code{12571} \code{12575} \code{12580} \code{12615} \code{12628} \code{12630} \code{12649} \code{12684} \code{12702} \code{12721} \code{12737} \code{12739} \code{12740} \code{12759} \code{12768} \code{12772} \code{12774} \code{12780} \code{12816} \code{12819} \code{12821} \code{12822} \code{12823} \code{12825} \code{12827} \code{12828} \code{12829} \code{12831} \code{12834} \code{12835} \code{12837} \code{12842} \code{12843} \code{12865} \code{12866} \code{12867} \code{12870} \code{12876} \code{12877} \code{12931} \code{12953} \code{12977} \code{12983} \code{13003} \code{13014} \code{13024} \code{13038} \code{13039} \code{13040} \code{13051} \code{13056} \code{13058} \code{13063} \code{13070} \code{13086} \code{13105} \code{13106} \code{13109} \code{13118} \code{13120} \code{13121} \code{13123} \code{13132} \code{13136} \code{13142} \code{13170} \code{13171} \code{13179} \code{13180} \code{13195} \code{13197} \code{13202} \code{13214} \code{13361} \code{13382} \code{13419} \code{13479} \code{13482} \code{13555} \code{13587} \code{13595} \code{13602} \code{13614} \code{13617} \code{13645} \code{13653} \code{13654} \code{13710} \code{13730} \code{13731} \code{13804} \code{13838} \code{13850} \code{13853} \code{13860} \code{13870} \code{13876} \code{13897} \code{14004} \code{14017} \code{140474} \code{14056} \code{14057} \code{140577} \code{14060} \code{14062} \code{14063} \code{14066} \code{14067} \code{14073} \code{14081} \code{14089} \code{14115} \code{14118} \code{14120} \code{14121} \code{14127} \code{14129} \code{14131} \code{14148} \code{14164} \code{14165} \code{14187} \code{14190} \code{14194} \code{14200} \code{14235} \code{14251} \code{14252} \code{14254} \code{14263} \code{14268} \code{142681} \code{14275} \code{14284} \code{14302} \code{14313} \code{14314} \code{14317} \code{14325} \code{14343} \code{14348} \code{14349} \code{14373} \code{14377} \code{14381} \code{14390} \code{14396} \code{14402} \code{14411} \code{14412} \code{14420} \code{14421} \code{14431} \code{14433} \code{14453} \code{14468} \code{14469} \code{14473} \code{14527} \code{14537} \code{14555} \code{14561} \code{14571} \code{14579} \code{14585} \code{14598} \code{14621} \code{14629} \code{14678} \code{14708} \code{14711} \code{14725} \code{14726} \code{14734} \code{14735} \code{14744} \code{14758} \code{14775} \code{14778} \code{14825} \code{14854} \code{14874} \code{14960} \code{14972} \code{14979} \code{14999} \code{15007} \code{15015} \code{15122} \code{15162} \code{15163} \code{15201} \code{15211} \code{15223} \code{15233} \code{15273} \code{15277} \code{15353} \code{15356} \code{15361} \code{15372} \code{15410} \code{15415} \code{15431} \code{15438} \code{15445} \code{15446} \code{15451} \code{15458} \code{15460} \code{15486} \code{15493} \code{15495} \code{15502} \code{15507} \code{15510} \code{15519} \code{15894} \code{15901} \code{15903} \code{15929} \code{15931} \code{15932} \code{15937} \code{15945} \code{15958} \code{15979} \code{15980} \code{16000} \code{16006} \code{16009} \code{16011} \code{16154} \code{16176} \code{16181} \code{16194} \code{16196} \code{16211} \code{16322} \code{16323} \code{16332} \code{16362} \code{16364} \code{16371} \code{16402} \code{16409} \code{16411} \code{16420} \code{16421} \code{16439} \code{16440} \code{16453} \code{16476} \code{16477} \code{16483} \code{16510} \code{16516} \code{16517} \code{16523} \code{16541} \code{16542} \code{16548} \code{16551} \code{16571} \code{16591} \code{16600} \code{16612} \code{16651} \code{16668} \code{16669} \code{16691} \code{16706} \code{16776} \code{16780} \code{16784} \code{16790} \code{16792} \code{16803} \code{16819} \code{16825} \code{16852} \code{16854} \code{16869} \code{16878} \code{16881} \code{16897} \code{16918} \code{16949} \code{16956} \code{16971} \code{16985} \code{16997} \code{17022} \code{170442} \code{170484} \code{17063} \code{170638} \code{170736} \code{170742} \code{170744} \code{170752} \code{170761} \code{17082} \code{17084} \code{17105} \code{17110} \code{17112} \code{171166} \code{17117} \code{17118} \code{171180} \code{171199} \code{17121} \code{171267} \code{171283} \code{171284} \code{17132} \code{171463} \code{17152} \code{17155} \code{17160} \code{17174} \code{17196} \code{17210} \code{17215} \code{17217} \code{17218} \code{17219} \code{17220} \code{17235} \code{17279} \code{17287} \code{17288} \code{17289} \code{17306} \code{17313} \code{17345} \code{17357} \code{17386} \code{17390} \code{17393} \code{17434} \code{17436} \code{17449} \code{17470} \code{17524} \code{17533} \code{17534} \code{17536} \code{17540} \code{17690} \code{17691} \code{17698} \code{17748} \code{17750} \code{17761} \code{17762} \code{17768} \code{17777} \code{17829} \code{17864} \code{17868} \code{17869} \code{17886} \code{17916} \code{17918} \code{17921} \code{17967} \code{17969} \code{17988} \code{17992} \code{17996} \code{18000} \code{18010} \code{18032} \code{18033} \code{18037} \code{18049} \code{18053} \code{18073} \code{18140} \code{18143} \code{18148} \code{18164} \code{18169} \code{18176} \code{18194} \code{18214} \code{18245} \code{18263} \code{18293} \code{18295} \code{18400} \code{18414} \code{18439} \code{18476} \code{18478} \code{18491} \code{18553} \code{18573} \code{18595} \code{18596} \code{18597} \code{18604} \code{18606} \code{18611} \code{18627} \code{18628} \code{18636} \code{18703} \code{18707} \code{18718} \code{18720} \code{18722} \code{18726} \code{18736} \code{18777} \code{18787} \code{18793} \code{18799} \code{18806} \code{18810} \code{18817} \code{18824} \code{18830} \code{18971} \code{19015} \code{19017} \code{19027} \code{19039} \code{19042} \code{19073} \code{19123} \code{19125} \code{19130} \code{19143} \code{19200} \code{19212} \code{19215} \code{192166} \code{192176} \code{192192} \code{192201} \code{19228} \code{19240} \code{19241} \code{19260} \code{19261} \code{192653} \code{192897} \code{19293} \code{19298} \code{19299} \code{19301} \code{19329} \code{19362} \code{19363} \code{19378} \code{194352} \code{194401} \code{194590} \code{19652} \code{19657} \code{19663} \code{19679} \code{19691} \code{19731} \code{19775} \code{19876} \code{19935} \code{20042} \code{20103} \code{20104} \code{20130} \code{20132} \code{20133} \code{20135} \code{20148} \code{20190} \code{20195} \code{20198} \code{20200} \code{20201} \code{20216} \code{20259} \code{20266} \code{20272} \code{20280} \code{20288} \code{20293} \code{20295} \code{20296} \code{20297} \code{20304} \code{20305} \code{20306} \code{20308} \code{20310} \code{20311} \code{20315} \code{20319} \code{20322} \code{20341} \code{20345} \code{20356} \code{20358} \code{20378} \code{20410} \code{20440} \code{20442} \code{20449} \code{20450} \code{20465} \code{20501} \code{20504} \code{20505} \code{20510} \code{20515} \code{20518} \code{20529} \code{20531} \code{20532} \code{20533} \code{20539} \code{20586} \code{20598} \code{20655} \code{20666} \code{20677} \code{20681} \code{20682} \code{20692} \code{20700} \code{20708} \code{207151} \code{20716} \code{20720} \code{207212} \code{207215} \code{207227} \code{20723} \code{20750} \code{20768} \code{207683} \code{20776} \code{20778} \code{207818} \code{208194} \code{208228} \code{20848} \code{208628} \code{208665} \code{20877} \code{208777} \code{208890} \code{20893} \code{20897} \code{209003} \code{20916} \code{20917} \code{209176} \code{20933} \code{20937} \code{209446} \code{209584} \code{20969} \code{20970} \code{209737} \code{20977} \code{209773} \code{20980} \code{210293} \code{210463} \code{211323} \code{211389} \code{211401} \code{211468} \code{211623} \code{211651} \code{211739} \code{211770} \code{211961} \code{212427} \code{212439} \code{212442} \code{212516} \code{212531} \code{212627} \code{212647} \code{212712} \code{212862} \code{212933} \code{212937} \code{21336} \code{213417} \code{21346} \code{213539} \code{21354} \code{21355} \code{21356} \code{21401} \code{21405} \code{21412} \code{214133} \code{21452} \code{214579} \code{214642} \code{214791} \code{215051} \code{215085} \code{215210} \code{215335} \code{215387} \code{215632} \code{215772} \code{215819} \code{216049} \code{216134} \code{216161} \code{216197} \code{216225} \code{216285} \code{216459} \code{216616} \code{216643} \code{216767} \code{21678} \code{21682} \code{216835} \code{216860} \code{216864} \code{216892} \code{217262} \code{217333} \code{21743} \code{217517} \code{217578} \code{217593} \code{217664} \code{217666} \code{21769} \code{21781} \code{217847} \code{217864} \code{217869} \code{21807} \code{21808} \code{21810} \code{218138} \code{21816} \code{21817} \code{21825} \code{218341} \code{218454} \code{218476} \code{21857} \code{21858} \code{21859} \code{218820} \code{218973} \code{21898} \code{219148} \code{219170} \code{219189} \code{21923} \code{21926} \code{21929} \code{21934} \code{21937} \code{21938} \code{21973} \code{21982} \code{21983} \code{22003} \code{22017} \code{22042} \code{22059} \code{22061} \code{22063} \code{22117} \code{22137} \code{22139} \code{22142} \code{22143} \code{22151} \code{22154} \code{22177} \code{22192} \code{22194} \code{22228} \code{22239} \code{22240} \code{22249} \code{22253} \code{22256} \code{22288} \code{22323} \code{22329} \code{22333} \code{223337} \code{22337} \code{22339} \code{223453} \code{223455} \code{22351} \code{22352} \code{22361} \code{223726} \code{223775} \code{22379} \code{223918} \code{224024} \code{22404} \code{224088} \code{224109} \code{224405} \code{224697} \code{225182} \code{225187} \code{225215} \code{225288} \code{225600} \code{225908} \code{22599} \code{22608} \code{226098} \code{226101} \code{226169} \code{22646} \code{226610} \code{226646} \code{226654} \code{226695} \code{22697} \code{226999} \code{227095} \code{227197} \code{22751} \code{227541} \code{227632} \code{227699} \code{227720} \code{22776} \code{22782} \code{22793} \code{228061} \code{228357} \code{228413} \code{228432} \code{228482} \code{228543} \code{228607} \code{228608} \code{228677} \code{228852} \code{228911} \code{229302} \code{229323} \code{229697} \code{229699} \code{229731} \code{229759} \code{229841} \code{230098} \code{230126} \code{230163} \code{230459} \code{230514} \code{230558} \code{230579} \code{230584} \code{230752} \code{230787} \code{230810} \code{230872} \code{231050} \code{231070} \code{231086} \code{231130} \code{231238} \code{231507} \code{231532} \code{231633} \code{231805} \code{231991} \code{232078} \code{232313} \code{232345} \code{232406} \code{232441} \code{232493} \code{232714} \code{232970} \code{233038} \code{233199} \code{233274} \code{233328} \code{233406} \code{233571} \code{233744} \code{233799} \code{233836} \code{234076} \code{234214} \code{234267} \code{234353} \code{234356} \code{234396} \code{234564} \code{234671} \code{234686} \code{234725} \code{234839} \code{234875} \code{234911} \code{235043} \code{235086} \code{235130} \code{235300} \code{235312} \code{235386} \code{235493} \code{235582} \code{235627} \code{235636} \code{236149} \code{236266} \code{236285} \code{236792} \code{236794} \code{236900} \code{236930} \code{237082} \code{237320} \code{237339} \code{237387} \code{237436} \code{237542} \code{237860} \code{238317} \code{23834} \code{23836} \code{238393} \code{23849} \code{23850} \code{23876} \code{23886} \code{23887} \code{23888} \code{238988} \code{239096} \code{239133} \code{23918} \code{23923} \code{239273} \code{239318} \code{23943} \code{23945} \code{23956} \code{23957} \code{23959} \code{239606} \code{239667} \code{239691} \code{239827} \code{23985} \code{239857} \code{240057} \code{240549} \code{24059} \code{24067} \code{240672} \code{240675} \code{24088} \code{240888} \code{240899} \code{24108} \code{24136} \code{241431} \code{241612} \code{241794} \code{241877} \code{242050} \code{242122} \code{242248} \code{242297} \code{242521} \code{243168} \code{243385} \code{243743} \code{243771} \code{244049} \code{244448} \code{244579} \code{244650} \code{244668} \code{244757} \code{244853} \code{244871} \code{245038} \code{245643} \code{245650} \code{245666} \code{245886} \code{245945} \code{246104} \code{246277} \code{246316} \code{246691} \code{246694} \code{246700} \code{246707} \code{246728} \code{246738} \code{252864} \code{257632} \code{257662} \code{258591} \code{259172} \code{259301} \code{260297} \code{260299} \code{26357} \code{26362} \code{26367} \code{26377} \code{26379} \code{26382} \code{26388} \code{26401} \code{26432} \code{26456} \code{26458} \code{26462} \code{26558} \code{26564} \code{266645} \code{268396} \code{268451} \code{268697} \code{268709} \code{268782} \code{268860} \code{26888} \code{268930} \code{26897} \code{268970} \code{26903} \code{269132} \code{269209} \code{269224} \code{269261} \code{26934} \code{26940} \code{269437} \code{26949} \code{269582} \code{26968} \code{269799} \code{269831} \code{269855} \code{26992} \code{270076} \code{270166} \code{27045} \code{27052} \code{27053} \code{270685} \code{271036} \code{271047} \code{271127} \code{271711} \code{271849} \code{27204} \code{27205} \code{27219} \code{27226} \code{27273} \code{27279} \code{27360} \code{27361} \code{27364} \code{27369} \code{27371} \code{27384} \code{27425} \code{276950} \code{277753} \code{279572} \code{27973} \code{27984} \code{28028} \code{28030} \code{280645} \code{28193} \code{28248} \code{28254} \code{286940} \code{29807} \code{29811} \code{29815} \code{29818} \code{30045} \code{30058} \code{30791} \code{30805} \code{30877} \code{30878} \code{30935} \code{30938} \code{30954} \code{30955} \code{30956} \code{319176} \code{319177} \code{319180} \code{319184} \code{319186} \code{319530} \code{319554} \code{319634} \code{319710} \code{319800} \code{319804} \code{319832} \code{319876} \code{319880} \code{319901} \code{319939} \code{320214} \code{320343} \code{320415} \code{320435} \code{320484} \code{320506} \code{320825} \code{320832} \code{320981} \code{320982} \code{327766} \code{328092} \code{328417} \code{329152} \code{329154} \code{329384} \code{329421} \code{329650} \code{329872} \code{330050} \code{330097} \code{330189} \code{330485} \code{330695} \code{330812} \code{330836} \code{330962} \code{331487} \code{331535} \code{333050} \code{333605} \code{338348} \code{338367} \code{338521} \code{353156} \code{353172} \code{353211} \code{360213} \code{380728} \code{380845} \code{380882} \code{380997} \code{381293} \code{381484} \code{381489} \code{381511} \code{381801} \code{382051} \code{382117} \code{382793} \code{387524} \code{432478} \code{432769} \code{433667} \code{434203} \code{434437} \code{434674} \code{50529} \code{50530} \code{50708} \code{50720} \code{50762} \code{50770} \code{50781} \code{50789} \code{50790} \code{50791} \code{50905} \code{50909} \code{50934} \code{51788} \code{51793} \code{51797} \code{51798} \code{51902} \code{51938} \code{51944} \code{52014} \code{52033} \code{52118} \code{52250} \code{52276} \code{52469} \code{52502} \code{52530} \code{52637} \code{52666} \code{52685} \code{52815} \code{52846} \code{52898} \code{53315} \code{53330} \code{53376} \code{53415} \code{53418} \code{53422} \code{53859} \code{53881} \code{54124} \code{54135} \code{54140} \code{54208} \code{54392} \code{54397} \code{54403} \code{54405} \code{54419} \code{54420} \code{54445} \code{54448} \code{545030} \code{54525} \code{545389} \code{545893} \code{54601} \code{54611} \code{54613} \code{54646} \code{54648} \code{54683} \code{547253} \code{55932} \code{55938} \code{55942} \code{55951} \code{55960} \code{55980} \code{55983} \code{56016} \code{56018} \code{56045} \code{56066} \code{56078} \code{56185} \code{56193} \code{56200} \code{56233} \code{56264} \code{56375} \code{56405} \code{56434} \code{56451} \code{56452} \code{56470} \code{56484} \code{56489} \code{56532} \code{56550} \code{56620} \code{56695} \code{56696} \code{56699} \code{56708} \code{56722} \code{56727} \code{56752} \code{56753} \code{56808} \code{56838} \code{57278} \code{57319} \code{57738} \code{57741} \code{57754} \code{57781} \code{57916} \code{58185} \code{58186} \code{58187} \code{58210} \code{58234} \code{58243} \code{58250} \code{58810} \code{58859} \code{58861} \code{58866} \code{58875} \code{58909} \code{59020} \code{59049} \code{59126} \code{595136} \code{60363} \code{60525} \code{60599} \code{623230} \code{624866} \code{626870} \code{628900} \code{629147} \code{632971} \code{634731} \code{636808} \code{637021} \code{63828} \code{63953} \code{64050} \code{64051} \code{64074} \code{64085} \code{64177} \code{64293} \code{64339} \code{64380} \code{64384} \code{64406} \code{64435} \code{64452} \code{64659} \code{64697} \code{64898} \code{64918} \code{65086} \code{65112} \code{65221} \code{654795} \code{654812} \code{65962} \code{65969} \code{66046} \code{66049} \code{66052} \code{66060} \code{66065} \code{66071} \code{66083} \code{66092} \code{66109} \code{66142} \code{66153} \code{66154} \code{66166} \code{66174} \code{66249} \code{66254} \code{66269} \code{66302} \code{66333} \code{66377} \code{66387} \code{66412} \code{66445} \code{66480} \code{664883} \code{66495} \code{664994} \code{66548} \code{66566} \code{66594} \code{66601} \code{666048} \code{666348} \code{666926} \code{66694} \code{66696} \code{66717} \code{66725} \code{667277} \code{66729} \code{66809} \code{668173} \code{66821} \code{668218} \code{668225} \code{668303} \code{66834} \code{66848} \code{66885} \code{66890} \code{66895} \code{66904} \code{66945} \code{66948} \code{66949} \code{66950} \code{66953} \code{66977} \code{66988} \code{67003} \code{67017} \code{67041} \code{67044} \code{67092} \code{67111} \code{67115} \code{67130} \code{67131} \code{67133} \code{67153} \code{67169} \code{67171} \code{67182} \code{67184} \code{67220} \code{67230} \code{67279} \code{67299} \code{67338} \code{67399} \code{67420} \code{67441} \code{67442} \code{67457} \code{67460} \code{67464} \code{67472} \code{67473} \code{67547} \code{67579} \code{67606} \code{67622} \code{67695} \code{67701} \code{67713} \code{67732} \code{67742} \code{67758} \code{67760} \code{67776} \code{67800} \code{67801} \code{67811} \code{67847} \code{67864} \code{67880} \code{67973} \code{67997} \code{68014} \code{68021} \code{68024} \code{68027} \code{68036} \code{68044} \code{68180} \code{68194} \code{68197} \code{68294} \code{68316} \code{68339} \code{68344} \code{68349} \code{68352} \code{68371} \code{68436} \code{68453} \code{68493} \code{68526} \code{68576} \code{68585} \code{68628} \code{68631} \code{68671} \code{68680} \code{68709} \code{68724} \code{68738} \code{68743} \code{68774} \code{68775} \code{68777} \code{68816} \code{68836} \code{68917} \code{68943} \code{69090} \code{69121} \code{69123} \code{69219} \code{69263} \code{69288} \code{69299} \code{69379} \code{69386} \code{69387} \code{69524} \code{69574} \code{69581} \code{69698} \code{69719} \code{69742} \code{69754} \code{69769} \code{69772} \code{69774} \code{69836} \code{69875} \code{70021} \code{70026} \code{70059} \code{70101} \code{70120} \code{70127} \code{70152} \code{70155} \code{70218} \code{70292} \code{70296} \code{70333} \code{70337} \code{70385} \code{70397} \code{70456} \code{70466} \code{70584} \code{70615} \code{70652} \code{70673} \code{70784} \code{70797} \code{70911} \code{70950} \code{71059} \code{71091} \code{71279} \code{71306} \code{71325} \code{71367} \code{71474} \code{71481} \code{71538} \code{71562} \code{71602} \code{71660} \code{71670} \code{71682} \code{71684} \code{71685} \code{71687} \code{71690} \code{71707} \code{71733} \code{71743} \code{71764} \code{71781} \code{71790} \code{71827} \code{71900} \code{71904} \code{71911} \code{71927} \code{71934} \code{71963} \code{71973} \code{71982} \code{71994} \code{72003} \code{72090} \code{72119} \code{72269} \code{72290} \code{72293} \code{72350} \code{72391} \code{72486} \code{72549} \code{72607} \code{72640} \code{72736} \code{72899} \code{72900} \code{72935} \code{72972} \code{72973} \code{73075} \code{73078} \code{73132} \code{73149} \code{73230} \code{73247} \code{73251} \code{73296} \code{73569} \code{73608} \code{73656} \code{73710} \code{73723} \code{73804} \code{73822} \code{73914} \code{74008} \code{74012} \code{74020} \code{74039} \code{74087} \code{74091} \code{74114} \code{74122} \code{74129} \code{74144} \code{74145} \code{74147} \code{74156} \code{74159} \code{74186} \code{74194} \code{74202} \code{74241} \code{74245} \code{74337} \code{74338} \code{74340} \code{74407} \code{74480} \code{74521} \code{74568} \code{74616} \code{74646} \code{74754} \code{74761} \code{74776} \code{74901} \code{75033} \code{75050} \code{75104} \code{75296} \code{75400} \code{75404} \code{75406} \code{75512} \code{75547} \code{75552} \code{75572} \code{75578} \code{75600} \code{75605} \code{75712} \code{75736} \code{75750} \code{75909} \code{76044} \code{76117} \code{76131} \code{76137} \code{76187} \code{76206} \code{76238} \code{76263} \code{76279} \code{76281} \code{76332} \code{76408} \code{76469} \code{76507} \code{76527} \code{76551} \code{76589} \code{76608} \code{76763} \code{76808} \code{76843} \code{76846} \code{76900} \code{76933} \code{76959} \code{77031} \code{77318} \code{77337} \code{77480} \code{77552} \code{77583} \code{77683} \code{77794} \code{77864} \code{77889} \code{77963} \code{78004} \code{78242} \code{78339} \code{78388} \code{78558} \code{78634} \code{78754} \code{78781} \code{78784} \code{78785} \code{78920} \code{79201} \code{80281} \code{80334} \code{80707} \code{80721} \code{80782} \code{80837} \code{80859} \code{80888} \code{80891} \code{80907} \code{80911} \code{81799} \code{81896} \code{81897} \code{81907} \code{83397} \code{83436} \code{83453} \code{83671} \code{83796} \code{83813} \code{83924} \code{84094} \code{84585} \code{93686} \code{93694} \code{93695} \code{93721} \code{93732} \code{93734} \code{93747} \code{93761} \code{93834} \code{93842} \code{93890} \code{94071} \code{94092} \code{94176} \code{94212} \code{94224} \code{94229} \code{94352} \code{97130} \code{98256} \code{98314} \code{98363} \code{98402} \code{98432} \code{98496} \code{98878} \code{99571} \code{99887}}
\item{\code{TargetScan}}{a numeric vector}
\item{\code{miRDB}}{a numeric vector}
\item{\code{Predicted_Interactions}}{a numeric vector}
\item{\code{miRTarBase}}{a numeric vector}
\item{\code{Pred_Fun}}{a numeric vector}
}
}
\details{
Dataframe of matrix_mining after DataMiningMatrix function. To use for 
vignette building.
}
\source{
TimiRGeN package.
}
\references{
NA
}
\examples{
data(Int_data)
## maybe str(Int_data) ; plot(Int_data) ...
}
\keyword{datasets}
