% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMapp.R
\name{MakeMapp}
\alias{MakeMapp}
\title{MakeMapp}
\usage{
MakeMapp(filt_df, miR_IDs_adj, Datatype = '')
}
\arguments{
\item{filt_df}{Dataframe of mined microRNA-mRNA interactions}

\item{miR_IDs_adj}{miR_enseml_adj or miR_entrez_adj}

\item{Datatype}{L for entrez or En for ensembl}
}
\value{
A dataframe which should be saved as a text file for import into
pathvisio.
}
\description{
Create input for the MAPP plugin in pathvisio.
}
\examples{
library(org.Mm.eg.db)
library(clusterProfiler)
mm_miR -> miR
getIDs_miR_mouse(miR)
Int_matrix <- data.frame(row.names = c("mmu-miR-320-3p:Acss1",
"mmu-miR-27a-3p:Odc1"),
avecor = c(-0.9191653, 0.7826041),
miR = c("mmu-miR-320-3p", "mmu-miR-27a-3p"),
mRNA = c("Acss1", "Acss1"),
miR_Entrez = c(NA, NA),
mRNA_Entrez = c(68738, 18263),
TargetScan = c(1, 0),
miRDB = c(0, 0),
Predicted_Interactions = c(1, 0),
miRTarBase = c(0, 1),
Pred_Fun = c(1, 1))
MakeMapp(filt_df = Int_matrix, miR_IDs_adj = miR_adjusted_entrez,
Datatype = 'L')
}
