\name{MfuzzData}
\alias{MfuzzData}
\docType{data}
\title{
MfuzzData
}
\description{
MfuzzData from CreateClusters function in TimiRGeN. For use in passing
vignette build process on bioconductor.
}
\usage{data("MfuzzData")}
\format{
Expressionset object.
}
\examples{
data(MfuzzData)
## maybe str(MfuzzData) ; plot(MfuzzData) ...
}
\keyword{datasets}
