% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroRNA_full.R
\name{MicroRNA_full}
\alias{MicroRNA_full}
\title{MicroRNA_full}
\usage{
MicroRNA_full(miRdf, species)
}
\arguments{
\item{miRdf}{miR dataframe.}

\item{species}{Species e.g. hsa or mmu.}
}
\value{
hsa-miR-140-5p would be changed to microRNA 140-1
}
\description{
Changes miR names into microRNA names in one step.
This has less flexibility that MicroRNA_partial.
}
\examples{
mm_miR -> miR
MicroRNA_full(rownames(miR), species = "mmu") -> microRNA_genenames
}
