% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micrornaFull.R
\name{micrornaFull}
\alias{micrornaFull}
\title{micrornaFull}
\usage{
micrornaFull(miRdf, species)
}
\arguments{
\item{miRdf}{miR dataframe.}

\item{species}{Species e.g. hsa or mmu.}
}
\value{
hsa-miR-140-5p would be changed to microRNA 140-1
}
\description{
Changes miR names into microRNA names in one step.
This has less flexibility that MicroRNA_partial.
}
\examples{
miR <- mm_miR
microRNA_genenames <- micrornaFull(rownames(miR), species = "mmu")
}
