% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRDB_data.R
\name{miRDB_data}
\alias{miRDB_data}
\title{miRDB_data}
\usage{
miRDB_data(MAE, mirdb, species = '', orgDB)
}
\arguments{
\item{MAE}{MultiAssayExperiment.}

\item{mirdb}{Most recent predicted targets data file. Download from here
http://mirdb.org/download.html.}

\item{species}{Species of interest e.g. hsa or mmu.}

\item{orgDB}{organism db file specific for species being looked into.}
}
\value{
A dataframe of mRNA-miR interactons, and other information
associated with these interactions.
}
\description{
Produces a list of mRNA-miR predicted interactions from miRDB
for the species of interest from miRDB. Requires the tidyverse package.
}
\examples{
library(MultiAssayExperiment)
library(org.Mm.eg.db)
library(tidyverse)
library(clusterProfiler)
MAE <- MultiAssayExperiment(list(miRDB_data = data.frame(
                         row.names = c("5151919", "5161682", "5151921",
                                       "5151922"),
                         "V1" = c("mmu-let-7a-1-3p", "mmu-let-7c-2-3p",
                          "mmu-let-7k","mmu-miR-100-3p"),
                         "V2" = c("NM_144958", "NM_178648", "NM_133355",
                          "NM_172405"),
                         "V3" = c(53.89190, 53.99680, 83.25980,64.69390)))) 
MAE <- miRDB_data(MAE, mirdb = assay(MAE, 1), species = 'mmu',
                        orgDB = org.Mm.eg.db)
}
