% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRTarBase_data.R
\name{miRTarBase_data}
\alias{miRTarBase_data}
\title{miRTarBase_data}
\usage{
miRTarBase_data(mirtarbase, species = '')
}
\arguments{
\item{mirtarbase}{Datafile of the most recent experimental mRNA-miR
interactions. Download from here
http://mirtarbase.mbc.nctu.edu.tw/php/download.php.}

\item{species}{Species being studies e.g. hsa or mmu.}
}
\value{
Dataframe of mRNA-miR interactions with functional evidence.
}
\description{
Produces a list of mRNA-miR interactions with functional
validation (weaker validation methods are not included) from miRDB for the
species of interest from targetscans.
}
\examples{
library(tidyverse)
data.frame(row.names = c("506668", "506669", "506670"),
"miRTarBase.ID" = c("MIRT000005", "MIRT000005", "MIRT000016"),
"miRNA" = c("mmu-miR-124-3p", "mmu-miR-124-3p", "mmu-miR-210-3p"),
"Species..miRNA." = c(rep("Mus musculus", 3)),
"Target.Gene" = c("Itgb1", "Itgb1", "Tcf7l2"),
"Target.Gene..Entrez.ID." = c("16412", "16412", "21416"),
"Species..Target.Gene." = c(rep("Mus musculus", 3)),
"Experiments" = c("Luciferase reporter assay//Microarray//qRT-PCR",
"Luciferase reporter assay//qRT-PCR//Western blot//Reporter assay;Microarray",
"Luciferase reporter assay//Western blot//Microarray"),
"Support.Type" = c(rep("Functional MTI", 3)),
"References..PMID." = c("18042700", "18619591",
"20492721")) -> miRTarBase
miRTarBase_data(mirtarbase = miRTarBase, species = 'mmu') -> miRTarBase_mmu
}
