% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirtarbaseData.R
\name{mirtarbaseData}
\alias{mirtarbaseData}
\title{mirtarbaseData}
\usage{
mirtarbaseData(MAE, mirtarbase, species = '')
}
\arguments{
\item{MAE}{Multiassayexperiment object.}

\item{mirtarbase}{Output from dloadMirtarbase. Download from here
http://mirtarbase.mbc.nctu.edu.tw/php/download.php.}

\item{species}{Species being studies e.g. hsa or mmu.}
}
\value{
Dataframe of mRNA-miR interactions with functional evidence.
}
\description{
Produces a list of mRNA-miR interactions with functional
validation from miRTarBase (weaker validation methods are not included).
}
\examples{
MAE <- MultiAssayExperiment(list("mirtarbase" = data.frame(
                            miRTarBase.ID = c("MIRT006298", "MIRT737375"),
                            miRNA = c("hsa-miR-200b-3p", "hsa-miR-1321"),
                            Species..miRNA. = c("Homo sapiens",
                                                "Homo sapiens"),
                            Target.Gene = c("E2F3", "STX1A"),
                            Target.Gene..Entrez.ID.= c(1871, 6804),
                            Species..Target.Gene.=c("Homo sapiens",
                                                     "Homo sapiens"),
                           Experiments = c(" Flow//Luciferase reporter assay//Western blot",
                                           "PAR-CLIP"),
                           Support.Type = c("Functional MTI",
                                            "Functional MTI (Weak)"),
                           References..PMID. = c(22144583, 26701625))))

MAE <- mirtarbaseData(MAE = MAE, mirtarbase = assay(MAE, 1),
                      species = "hsa")
}
