% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_unique.R
\name{non_unique}
\alias{non_unique}
\title{non_unique}
\usage{
non_unique(Col, sep, suffix)
}
\arguments{
\item{Col}{Column of dataframe to modify.}

\item{sep}{Seperator to insert.}

\item{suffix}{Suffix to add behind seperator, default is 1, and increases
in numerical order for each repeated string in the column.}
}
\value{
Modified column of a dataframe. Each non-unique value will have
numbers representing it's duplicate number behing it.
}
\description{
Internal function for getIDs_miR_mousetohuman. Helps data
wrangling for -3p and -5p genes.
}
\examples{
sample <- data.frame(name = c("mmu-miR-101a-3p", "mmu-miR-101a-5p", 
                             "mmu-miR-101c","mmu-miR-106a-5p", 
                             "mmu-miR-106b-3p"),
          Hs_n = c("hsa-miR-101-1", "hsa-miR-101-1", "hsa-miR-101c",
                   "hsa-miR-106a", "hsa-miR-106b"))
sample$Hs_n <- non_unique(Col = sample$Hs_n, sep = '-', suffix = 'p') 
}
