% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonUnique.R
\name{nonUnique}
\alias{nonUnique}
\title{nonUnique}
\usage{
nonUnique(Col, sep, suffix)
}
\arguments{
\item{Col}{Column of dataframe to modify.}

\item{sep}{Seperator to insert.}

\item{suffix}{Suffix to add behind seperator, default is 1, and increases
in numerical order for each repeated string in the column.}
}
\value{
Modified column of a dataframe. Each non-unique value will have
numbers representing it's duplicate number behing it.
}
\description{
Internal function for getIDs functions. Helps data
wrangling for -3p and -5p genes.
}
