% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quickbar.R
\name{Quickbar}
\alias{Quickbar}
\title{Quickbar}
\usage{
Quickbar(X, Y)
}
\arguments{
\item{X}{Dataframe including count information and wikipathways.}

\item{Y}{Point to the associated name within a list of nested dataframes.}
}
\value{
barplot
}
\description{
Creates dotplot which compares the confidence level we have for
each wikipathways association to the data, the number of genes associated
to each wikipathway and this is all in a ranked fashion for each
timepoint and genetype.
}
\examples{
library(ggplot2)
library(org.Mm.eg.db)
library(clusterProfiler)
e_list -> elist
downloadGMT(speciesInitial = "Mm")
EnrichWiki(method = 'c', elist, org.Mm.eg.db,
path_gene, path_name, ID = 'ENTREZID',
universe = path_gene$gene) -> sigwiki
Quickbar(X = sigwiki$D7_wikipathways, Y = sigwiki[4]) -> Q
}
