% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quickfuzz.R
\name{Quickfuzz}
\alias{Quickfuzz}
\title{Quickfuzz}
\usage{
Quickfuzz(Mfuzzdata, Clusters, W)
}
\arguments{
\item{Mfuzzdata}{Output after CreateClusters. An Expressionset object.}

\item{Clusters}{Output after CreateClusters. A list.}

\item{W}{Should the plot be shown in a new window? Default is TRUE.}
}
\value{
A plot of different clusters showing how the input data varies
across different wikipathways.
}
\description{
Plots Clusters in reference to the ClusterData dataframe. To be
used after CreateClusters and ClusterCheck. From here individual
wikipathways that show variance throughout a signalling network can be
taken for further miR-mRNA interaction analysis.
}
\examples{
library(Mfuzz)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE
e_list -> MAE@metadata$elist
w_list[1:10] -> MAE@metadata$wlist
WikiMatrix(e_list = MAE@metadata$elist, 
wp_list = MAE@metadata$wlist) -> MAE@ExperimentList$Wmat

TurnPercent(wikiMatrix = MAE@ExperimentList$Wmat,
rowInt = 4) -> MAE@ExperimentList$Pmat

CreateClusters(method = "c", MAE =  MAE,
Percent_matrix = MAE@ExperimentList$Pmat,
no.clusters = 2, Variance = 0.99) -> MAE

Quickfuzz(Mfuzzdata = MAE@ExperimentList$MfuzzData,
Clusters = MAE@metadata$Clusters, W = FALSE)
}
