% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quicknet.R
\name{Quicknet}
\alias{Quicknet}
\title{Quicknet}
\usage{
Quicknet(net)
}
\arguments{
\item{net}{network generated by Makenet function. If you have fewer than two
interactions the plot will not work.}
}
\value{
A network depicting miR-mRNA interactions.
}
\description{
Generates an internal R network representing the miR-mRNA
interactions filtered through the input data using wikipathways analysis and
database filtering/
}
