% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickReg.R
\name{quickReg}
\alias{quickReg}
\title{quickReg}
\usage{
quickReg(reg_df, colselect)
}
\arguments{
\item{reg_df}{Matrix/assay which which was generated by the multiReg function.
Should be in the MAE used during the multiReg function.}

\item{colselect}{Integer which represents a column within reg_df containing
mRNA/ miRNA which the gene of interest (column 2) is to be contrasted agianst.}
}
\value{
Regression plot between a miRNA:mRNA pair.
}
\description{
quickReg will produce a regression plot between a miRNA:mRNA
pair. An abline wil be generated. Odds-ratio will be calculated from
the coefficient. Confident intervals at 95% will be caclulated from the
coefficient also. This is a useful method to identify the relationship
between a miRNA:mRNA pair over time. The higher the odds-ratio the greater
the correlation between the two time series. The smaller the 95% confidence
intervals range, the more confident we can be about plotting the abline.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR[1:100,]

mRNA <- mm_mRNA[1:200,]

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- getIdsMir(MAE, assay(MAE, 1), orgDB = org.Mm.eg.db, 'mmu')

MAE <- getIdsMrna(MAE, assay(MAE, 2), "useast", 'mmusculus')

MAE <- diffExpressRes(MAE, df = assay(MAE, 1), dataType = 'Log2FC',
                     genes_ID = assay(MAE, 3),
                      idColumn = 'GENENAME',
                      name = "miRNA_log2fc")

MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
                     genes_ID = assay(MAE, 7),
                     idColumn = 'GENENAME',
                     name = "mRNA_log2fc")

Filt_df <- data.frame(row.names = c("mmu-miR-145a-3p:Adamts15",
                                   "mmu-miR-146a-5p:Acy1"),
                     corr = c(-0.9191653, 0.7826041),
                     miR = c("mmu-miR-145a-3p", "mmu-miR-146a-5p"),
                     mRNA = c("Adamts15", "Acy1"),
                     miR_Entrez = c(387163, NA),
                     mRNA_Entrez = c(235130, 109652),
                     TargetScan = c(1, 0),
                     miRDB = c(0, 0),
                     Predicted_Interactions = c(1, 0),
                     miRTarBase = c(0, 1),
                     Pred_Fun = c(1, 1))

MAE <- matrixFilter(MAE, miningMatrix = Filt_df, negativeOnly = FALSE,
                   threshold = 1, predictedOnly = FALSE)

MAE <- multiReg(MAE = MAE, gene_interest = "Adamts15",
                mRNAreg =TRUE, filt_df=MAE[[11]], miRNA_exp=MAE[[9]],
                mRNA_exp=MAE[[10]])

model1 <- linearRegr(mreg = MAE[[12]], colselect =2, colpair =3)

summary(model1$regression)

quickReg(reg_df = MAE[[12]], colselect = 3)
}
