% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickTC.R
\name{quickTC}
\alias{quickTC}
\title{quickTC}
\usage{
quickTC(filt_df, pair, miRNA_exp, mRNA_exp, scale,Interpolation,
timecourse)
}
\arguments{
\item{filt_df}{Dataframe from the matrixFilter function.}

\item{pair}{Interger representing the pair to be explored.}

\item{miRNA_exp}{miRNA data from using the diffExpressRes function on miRNA
data.}

\item{mRNA_exp}{mRNA data from using the diffExpressRes function on miRNA
data}

\item{scale}{TRUE or FALSE. Should data be scales. Default is FALSE.}

\item{Interpolation}{TRUE or FALSE. Should the whole time course be
interpolated over by a smooth spline? Default is FALSE. This is most useful
for longer time courses.}

\item{timecourse}{If Iterpolation is TRUE, how many time points should be
interpolated over?}
}
\value{
Time course plot of selected pair.
}
\description{
Plots miRNA:mRNA pair over timecourse.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR[1:50,]

mRNA <- mm_mRNA[1:100,]

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- getIdsMir(MAE, assay(MAE, 1), orgDB = org.Mm.eg.db, 'mmu')

MAE <- getIdsMrna(MAE, assay(MAE, 2), "useast", 'mmusculus', orgDB = org.Mm.eg.db)

MAE <- diffExpressRes(MAE, df = assay(MAE, 1), dataType = 'Log2FC',
                     genes_ID = assay(MAE, 3),
                      idColumn = 'GENENAME',
                      name = "miRNA_log2fc")

MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
                     genes_ID = assay(MAE, 7),
                     idColumn = 'GENENAME',
                     name = "mRNA_log2fc")

Filt_df <- data.frame(row.names = c("mmu-miR-145a-3p:Adamts15",
                                   "mmu-miR-146a-5p:Acy1"),
                     corr = c(-0.9191653, 0.7826041),
                     miR = c("mmu-miR-145a-3p", "mmu-miR-146a-5p"),
                     mRNA = c("Adamts15", "Acy1"),
                     miR_Entrez = c(387163, NA),
                     mRNA_Entrez = c(235130, 109652),
                     TargetScan = c(1, 0),
                     miRDB = c(0, 0),
                     Predicted_Interactions = c(1, 0),
                     miRTarBase = c(0, 1),
                     Pred_Fun = c(1, 1))

MAE <- matrixFilter(MAE, miningMatrix = Filt_df, negativeOnly = FALSE,
                   threshold = 1, predictedOnly = FALSE)

quickTC(filt_df=MAE[[11]], pair=1, miRNA_exp=MAE[[9]],
        mRNA_exp=MAE[[10]], scale = FALSE)
}
