% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceWiki.R
\name{reduceWiki}
\alias{reduceWiki}
\title{reduceWiki}
\usage{
reduceWiki(MAE, path_data, stringWiki = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the results of reduceWiki.
It is recommended to use the same MAE which stores the data from dloadGmt/
gmtEnsembl.}

\item{path_data}{Dataframe with wikipathway IDs, gene IDs and pathway names
from either the dloadGmt or gmtEnsembl functions. These will be found as
assays within the MAE used in the dloadGmt or gmtEnsembl functions.}

\item{stringWiki}{Full name of the wikipathway of interest. Make sure to
spell this correctly. Each wikipathway can only be added once to the same MAE
object.}
}
\value{
A dataframe that only contains information about the wikipathway of
interest. Output will be stored as an assay in the input MAE.
}
\description{
Returns all gene IDs of a single wikipathway of interest. This
function is recommended to be used after a signalling pathway of interest
is found.
}
\examples{
MAE <- MultiAssayExperiment(list(path_data = data.frame(
                        "wpid" = c(rep("WP571", 6)),
                        "gene" = c(16175, 12370,26419, 19249, 19645, 18479),
                        "name" = c(rep("Fas pathway and Stress induction of HSP regulation",
                        6)))))

MAE <- reduceWiki(MAE, path_data = assay(MAE, 1),
                 stringWiki = 'Fas pathway and Stress induction of HSP regulation')
}
