% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartObject.R
\name{StartObject}
\alias{StartObject}
\title{StartObject}
\usage{
StartObject(miR, mRNA)
}
\arguments{
\item{miR}{microRNA dataframe/ matrix}

\item{mRNA}{mRNA dataframe/ matrix}
}
\value{
MultiAssayExperiment object
}
\description{
Creates a MultiAssayObject from miR and mRNA dataframes. This
will be the constant object used throughout TimiRGeN.
}
\examples{
library(MultiAssayExperiment)
miR <- mm_miR
mRNA <- mm_mRNA
MAE <- StartObject(miR = miR, mRNA = mRNA)
}
