% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetscanData.R
\name{targetscanData}
\alias{targetscanData}
\title{targetscanData}
\usage{
targetscanData(MAE, targetScan, species = '')
}
\arguments{
\item{MAE}{MultiAssayExperiments.}

\item{targetScan}{File of non-conserved miR-mRNA predicted targets.
Downloaded from here
http://www.targetscan.org/cgi-bin/targetscan/data_download.vert72.cgi.}

\item{species}{Species we are interested in e.g hsa or mmu.}
}
\value{
List of predicted mRNA-miR interactions and associated information.
}
\description{
Produces a list of mRNA-miR predicted interactions from miRDB
for the species of interest from targetscans. Requires the tidyverse
package.
}
\examples{
library(tidyverse)
library(MultiAssayExperiment)
MAE <- MultiAssayExperiment(list(TargetScans = data.frame(
                         row.names = c("215", "216", "286", "287"),
                        "Gene" = c("ENSG00000148584.10",
                                   "ENSG00000148584.10",
                                   "ENSG00000109576.9",
                                   "ENSG00000109576.9"),
                         "ID" =c("A1CF", "A1CF", "AADAT", "AADAT"),
                         "Gene.1" = c("ENST00000374001.2",
                                      "ENST00000374001.2",
                                      "ENST00000337664.4",
                                      "ENST00000337664.4"),
                        "Symbol" = c("10090", "10090", "10090", "10090"),
                       "Transcript" = c("mmu-miR-429-3p", "mmu-miR-374c-5p",
                                    "mmu-miR-182-5p","mmu-miR-340-5p"))))
TargetScans_df <- targetscanData(MAE, targetScan = assay(MAE, 1),
                                  species = 'mmu')
}
