% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnPercent.R
\name{turnPercent}
\alias{turnPercent}
\title{turnPercent}
\usage{
turnPercent(MAE, wikiMatrix)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the output from
turnPercent. It is recommended to use the MAE which stores output from the
wikiMatrix function.}

\item{wikiMatrix}{Numeric matrix of wikipathways and samples. This should be
stored as an assay within the MAE used in the wikiMatrix function.}
}
\value{
A percentage matrix which contrasts genes found in pathways and
samples. Output will be stored as an assay within the input MAE.
}
\description{
Genes found in common between the input data and each
pathway are normalised by percentages. This is to normalise for pathway size.
}
\examples{
MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list_mouse

metadata(MAE)[["w_list"]] <- w_list_mouse[1:10]

MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE, wikiMatrix = assay(MAE, 1))
}
