\name{w_list_human}
\alias{w_list_human}
\docType{data}
\title{
w_list_human
}
\description{
Wikilist of human wikipathways and ensembl gene IDs.
}
\usage{data("w_list_human")}
\format{
list
}
\details{
A Large list of each wikipathway and associated genes.
}
\source{
https://www.wikipathways.org/index.php/Download_Pathways
}
\references{
Denise N Slenter et al. “WikiPathways: a multifaceted pathway database b
ridging metabolomics to other omics research”.
In: Nucleic acids research 46.D1 (2017), pp. D661–D667.
}
\examples{
data(w_list_human)
## maybe str(w_list_human) ; plot(w_list_human) ...
}
\keyword{datasets}
