% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiList.R
\name{wikiList}
\alias{wikiList}
\title{wikiList}
\usage{
wikiList(MAE, stringSpecies = '', stringSymbol = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment object which the results of wikiList is
to be added. It is recommended to use the same MAE which stores data from
dloadGmt.}

\item{stringSpecies}{Species name to decide which wikipathways to download.
'Homo sapiens' to download human pathways or 'Mus musculus' to download
mouse pathways.}

\item{stringSymbol}{Type of gene code e.g. 'H' for human gene names
, 'En' for ensemble gene IDs , 'L' for entrezgene IDs.}
}
\value{
List of wikipathways and associated genes as a list of strings.
}
\description{
Provides a list of wikipathways and each gene found to be
within that wikipathway. This will be stored as metadata of a MAE. wikiList
will download a large amount of data from the wikipathways website so this
may take some time.
}
\examples{
\donttest{
MAE <- MultiAssayExperiment()

MAE <- wikiList(MAE, stringSpecies = 'Homo sapiens', stringSymbol = 'En')
}
}
