% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiMatrix.R
\name{wikiMatrix}
\alias{wikiMatrix}
\title{wikiMatrix}
\usage{
wikiMatrix(MAE, ID_list, wp_list)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the output of
wikiMatrix. It is recommended that the MAE object which stores output from
the wikiList function.}

\item{ID_list}{List of lists of entrez gene IDs or ensembl gene IDs stored
as strings. This is the output of the eNames function and is stored as
metadata in the MAE used in the eNames function.}

\item{wp_list}{List of lists containing wikipathways with entrez gene IDs or
ensembl IDs as strings. This is the output of the wikiList function and is
stored as metadata in the MAE used in the wikiList function.}
}
\value{
A matrix showing which genes are found in each time points for
each pathway. Output will be stored as an assay in the input MAE.
}
\description{
Creates a matrix which shows how many genes from the input mRNA
and miRNA data are found in each wikipathway, for a specific species.
}
\examples{
MAE <- MultiAssayExperiment()

metadata(MAE)[["ID_list"]] <- e_list_mouse

metadata(MAE)[["w_list"]] <- w_list_mouse[1:10]

MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])
}
