% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIDs.R
\name{AddIDs}
\alias{AddIDs}
\title{AddIDs}
\usage{
AddIDs(method, filtered_genelist, miR_IDs, mRNA_IDs)
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{filtered_genelist}{A list of nested dataframes if 'c' or A list of
lists with nested dataframes if 's'.}

\item{miR_IDs}{miR_ensembl or miR_entrez. Use getIDs function to acquire
this.}

\item{mRNA_IDs}{mRNA_ensembl or mRNA_entrez. Use getIDs function to acquire
this.}
}
\value{
list of dataframes with entrezIDs and genenames additional
as columns.
}
\description{
Adds entrez or ensembl IDs to the nested dataframes within
the filtered_genelist.
}
\examples{
library(clusterProfiler)
library(org.Mm.eg.db)
miR <- mm_miR
miR <- miR[1:100,]
mRNA <- mm_mRNA
mRNA <- mRNA[1:200,]
getIDs_miR_mouse(miR = miR)
getIDs_mRNA_mouse(mRNA = mRNA)
CombineGenes(miR_data = miR, mRNA_data = mRNA) -> genetic_data
GenesList(method = 'c', genetic_data = genetic_data,
timeString = 'D') -> genelist
SignificantVals(method = 'c', geneList = genelist, maxVal = 0.05,
stringVal = 'adjPVal') -> filtered_genelist
AddIDs(method = 'c', filtered_genelist = filtered_genelist,
miR_IDs = miR_entrez, mRNA_IDs = mRNA_entrez) -> gene_entrez
}
