% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIds.R
\name{addIds}
\alias{addIds}
\title{addIds}
\usage{
addIds(MAE, method, filtered_genelist, miR_IDs, mRNA_IDs)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{filtered_genelist}{A list of nested dataframes if 'c' or A list of
lists with nested dataframes if 's'.}

\item{miR_IDs}{miR_ensembl or miR_entrez. Use getIDs function to acquire
this.}

\item{mRNA_IDs}{mRNA_ensembl or mRNA_entrez. Use getIDs function to acquire
this.}
}
\value{
list of dataframes with entrezIDs and genenames additional
as columns which can be stored in metadata section of an MAE.
}
\description{
Adds entrez or ensembl IDs to the nested dataframes within
the filtered_genelist.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR
mRNA <- mm_mRNA
Data <- startObject(miR = miR, mRNA = mRNA)

Data <- combineGenes(MAE = Data, miR_data = assay(Data, 1),
                     mRNA_data = assay(Data, 2))

Data <- genesList(MAE = Data, method = 'c', genetic_data = assay(Data, 3),
                  timeString = 'D')

Data <- significantVals(MAE = Data, method = 'c',
                        geneList = metadata(Data)[[1]],
                        maxVal = 0.05, stringVal = "adjPVal")

Data <- addIds(MAE = Data, method = "c",
              filtered_genelist = metadata(Data)[[2]],
              miR_IDs = assay(Data, 3), mRNA_IDs = assay(Data, 3))
}
