% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPrefix.R
\name{addPrefix}
\alias{addPrefix}
\title{addPrefix}
\usage{
addPrefix(MAE, gene_df, prefixString = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment}

\item{gene_df}{mRNA or miR results from differential expression analysis
inside of a MAE.}

\item{prefixString}{miR_ or mRNA_}
}
\value{
mRNA or miR results now should have mRNA_ or miR_ at the start
of each column name.
}
\description{
Adds 'miR_' or 'mRNA_' prefix to colnames of dataframes. Can
also add any other prefix, if user has other datatypes to explore. Colnames
should end up in the following naming type: 'genetype_timepoint.datatype'.
}
\examples{
miR <- mm_miR

mRNA <- mm_mRNA

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- addPrefix(MAE = MAE, gene_df = assay(MAE, 1),
                 prefixString = "miR")

MAE <- addPrefix(MAE = MAE, gene_df = assay(MAE, 2),
                 prefixString = "mRNA")
}
