\name{c_matrix}
\alias{c_matrix}
\docType{data}
\title{
c_matrix
}
\description{
Matrix of cluster information for the vignette.
}
\usage{data("c_matrix")}
\format{
The format is:
num [1:202, 1:12] 0.06292 0.00064 0.03325 0.02371 0.02164 ...
- attr(*, "dimnames")=List of 2
..$ : chr [1:202] "WP1" "WP10" "WP103" "WP108" ...
..$ : chr [1:12] "1" "2" "3" "4" ...
}
\details{
Matrix of cluster to wikipathways information.
}
\source{
From data.
}
\references{
NA
}
\examples{
data(c_matrix)
}
\keyword{datasets}
