% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterCheck.R
\name{ClusterCheck}
\alias{ClusterCheck}
\title{ClusterCheck}
\usage{
ClusterCheck(Clusters, W)
}
\arguments{
\item{Clusters}{Outout from CreateClusters function. A list of clusters,
statistics and phenodata.}

\item{W}{Should the plot be shown in a new window? Default is TRUE.}
}
\value{
A PCAplot showing distance of clusters.
}
\description{
Put data from CreateClusters through this funciton to check
if an appropriate number of clusters have been created. The closer the
circles are to one another the more likely that they should belong
to the same cluster. Read more about Mfuzz for a greater understanding
https://bioconductor.org/packages/release/bioc/html/rWikiPathways.html.
}
\examples{
library(Mfuzz)
library(MultiAssayExperiment)
MAE <- Clusters
ClusterCheck(Clusters = metadata(MAE)[[1]], W = FALSE)
}
