% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCheck.R
\name{clusterCheck}
\alias{clusterCheck}
\title{clusterCheck}
\usage{
clusterCheck(Clusters, W)
}
\arguments{
\item{Clusters}{A large list of clusters, statistics and phenodata. This
will be stored as metadata within the MAE used in the createClusters
function.}

\item{W}{TRUE or FALSE. Should the plot be shown in a new window? Default is
FALSE.}
}
\value{
A PCAplot showing distance of clusters.
}
\description{
clusterCheck creates a PCA plot using functions from Mfuzz. This
will help to indicate if an appropriate number of clusters have been created.
The closer the circles are to one another the more likely that they should
belong to the same cluster. Read more about Mfuzz here
https://bioconductor.org/packages/release/bioc/html/rWikiPathways.html.
}
\examples{
MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list_mouse

metadata(MAE)[["w_list"]] <- w_list_mouse[1:10]

MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE,
                   wikiMatrix = assay(MAE, 1))

MAE <- createClusters(MAE, method = "c",
                    percentMatrix = assay(MAE, 2),
                    noClusters = 2, variance = 0.99)

clusterCheck(Clusters = metadata(MAE)[[3]], W = FALSE)
}
