\name{ClusterData}
\alias{ClusterData}
\docType{data}
\title{
ClusterData
}
\description{
Cluster data from CreateClusters in TimiRGeN. For passing vignette build
process in bioconductor.
}
\usage{data("ClusterData")}
\format{
  A data frame with 202 observations on the following 13 variables.
  \describe{
    \item{\samp{1}}{a numeric vector}
    \item{\samp{2}}{a numeric vector}
    \item{\samp{3}}{a numeric vector}
    \item{\samp{4}}{a numeric vector}
    \item{\samp{5}}{a numeric vector}
    \item{\samp{6}}{a numeric vector}
    \item{\samp{7}}{a numeric vector}
    \item{\samp{8}}{a numeric vector}
    \item{\samp{9}}{a numeric vector}
    \item{\samp{10}}{a numeric vector}
    \item{\samp{11}}{a numeric vector}
    \item{\samp{12}}{a numeric vector}
    \item{\code{Description}}{a character vector}
  }
}
\examples{
data(ClusterData)
## maybe str(ClusterData) ; plot(ClusterData) ...
}
\keyword{datasets}
