\name{Clusters}
\alias{Clusters}
\docType{data}
\title{
Clusters
}
\description{
Clusters from CreateClusters function in TimiRGeN. For use in passing
vignette build process on bioconductor.
}
\usage{data("Clusters")}
\format{
  The format is:
List of 7
}
\examples{
data(Clusters)
## maybe str(Clusters) ; plot(Clusters) ...
}
\keyword{datasets}
