\name{clusters}
\alias{clusters}
\docType{data}
\title{
clusters
}
\description{
Clusters from createClusters function in TimiRGeN. For use in passing
vignette build process on bioconductor.
}
\usage{data("clusters")}
\format{
  The format is:
List of 7
}
\examples{
data(clusters)
## maybe str(clusters) ; plot(clusters) ...
}
\keyword{datasets}
