% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGenes.R
\name{combineGenes}
\alias{combineGenes}
\title{combineGenes}
\usage{
combineGenes(MAE, miR_data, mRNA_data)
}
\arguments{
\item{MAE}{Input MAE which stores results from combineGenes. It is
recommended to use the MAE which was used in startObject.}

\item{miR_data}{microRNA dataframe. Rows should be genes, columns are DE
results and time point. This should be the stored as an assay within the
MAE used in the startObject function.}

\item{mRNA_data}{mRNA dataframe. Rows should be genes, columns are DE results
and time point. This should be the stored as an assay within the MAE used in
the startObject function.}
}
\value{
A dataframe with combined miR and mRNA data. Will be stored as an
assay in the input MAE.
}
\description{
Combines miR and mRNA data into one dataframe. Input columns
should be written as :timepoint.DifferentialExpressionResultType e.g.
D1.log2fc or H6.adjPval. Column names should be the same for miR and mRNA
data. If a more detailed explanation of column nomenclature is needed please
read the vignette.combineGenes is essential for combined analysis of miR-mRNA
data. If using separate analysis, there is no need to use combineGenes.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

mRNA <- mm_mRNA

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- combineGenes(MAE = MAE, miR_data = assay(MAE, 1),
                    mRNA_data = assay(MAE, 2))
}
