% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineGenes.R
\name{CombineGenes}
\alias{CombineGenes}
\title{CombineGenes}
\usage{
CombineGenes(miR_data, mRNA_data)
}
\arguments{
\item{miR_data}{microRNA dataframe}

\item{mRNA_data}{mRNA dataframe}
}
\value{
A dataframe with combines miR and mRNA data which can be stored 
within the MAE object.
}
\description{
Combines miR and mRNA data into one dataframe. Input columns
should be written as follows:
timepoint.DEresult e.g. D1.log2fc or H6.adjPval.
}
\examples{
library(clusterProfiler)
library(org.Mm.eg.db)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE
CombineGenes(miR_data = MAE@ExperimentList$miR, mRNA_data = 
MAE@ExperimentList$mRNA) -> MAE@ExperimentList$genetic_data
}
