% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGenes.R
\name{combineGenes}
\alias{combineGenes}
\title{combineGenes}
\usage{
combineGenes(MAE, miR_data, mRNA_data)
}
\arguments{
\item{MAE}{MultiAssayExperiment}

\item{miR_data}{microRNA dataframe}

\item{mRNA_data}{mRNA dataframe}
}
\value{
A dataframe with combines miR and mRNA data which can be stored
within the MAE object.
}
\description{
Combines miR and mRNA data into one dataframe. Input columns
should be written as follows:
timepoint.DEresult e.g. D1.log2fc or H6.adjPval.
}
\examples{
library(clusterProfiler)
library(org.Mm.eg.db)
library(MultiAssayExperiment)
miR <- mm_miR
mRNA <- mm_mRNA
MAE <- startObject(miR = miR, mRNA = mRNA)
MAE <- combineGenes(MAE = MAE, miR_data = assay(MAE, 1),
                    mRNA_data = assay(MAE, 2))
}
