% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateClusters.R
\name{CreateClusters}
\alias{CreateClusters}
\title{CreateClusters}
\usage{
CreateClusters(MAE, method, percentMatrix, noClusters, dataString = '',
               variance)
}
\arguments{
\item{MAE}{MultiAssayObject where the cluster information will be stored.}

\item{method}{Either C or S for combined or separate analysis.}

\item{percentMatrix}{Output after TurnPercent.}

\item{noClusters}{Number of clusters to create, the default is 5.}

\item{dataString}{Only for use in S analysis. Insert the prefix string e.g.
mRNA or miR.}

\item{variance}{Numeric from 0-1 to control strictness of filtering. Higher
Variance means wikipathways that are undergoing a higher change in the
data will be clustered.}
}
\value{
Clusters(metadata): A list to be used as the input in plot functions.
Mfuzzdata: An input for QuickFuzz. Mfuzzdata(ExperimentList): An
ExpressionSet objectto be input for Mfuzz clustering. 
ClusterData(ExperimentList): A breakdown of how each pathway fitted with 
each cluster.
}
\description{
Create soft clusters to assess how your datasets change in
gene abundance during your time course in different wikipathways.
This function is to be used before ClusterCheck and Quickfuzz.
}
\examples{
library(MultiAssayExperiment)
library(Mfuzz)
MAE <- MultiAssayExperiment()
metadata(MAE)[["e_list"]] <- e_list
metadata(MAE)[["w_list"]] <- w_list[1:10]
MAE <- WikiMatrix(MAE, ID_list = metadata(MAE)[[1]], 
                  wp_list = metadata(MAE)[[2]])
 
MAE <- TurnPercent(MAE = MAE, 
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)

MAE <- CreateClusters(MAE, method = "c", 
                    percentMatrix = assay(MAE, 2),
                    noClusters = 2, variance = 0.99)
}
