% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateClusters.R
\name{CreateClusters}
\alias{CreateClusters}
\title{CreateClusters}
\usage{
CreateClusters(method, MAE, Percent_matrix, no.clusters, Data_string = '',
Variance)
}
\arguments{
\item{method}{Either C or S for combined or separate analysis.}

\item{MAE}{MultiAssayObject where the cluster information will be stored.}

\item{Percent_matrix}{Output after TurnPercent.}

\item{no.clusters}{Number of clusters to create, the default is 5.}

\item{Data_string}{Only for use in S analysis. Insert the prefix string e.g.
mRNA or miR.}

\item{Variance}{Numeric from 0-1 to control strictness of filtering. Higher
Variance means wikipathways that are undergoing a higher change in the
data will be clustered.}
}
\value{
Clusters(metadata): A list to be used as the input in plot functions.
Mfuzzdata: An input for QuickFuzz. Mfuzzdata(ExperimentList): An
ExpressionSet objectto be input for Mfuzz clustering. 
ClusterData(ExperimentList): A breakdown of how each pathway fitted with 
each cluster.
}
\description{
Create soft clusters to assess how your datasets change in
gene abundance during your time course in different wikipathways.
This function is to be used before ClusterCheck and Quickfuzz.
}
\examples{
library(Mfuzz)
library(rWikiPathways)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE
e_list -> MAE@metadata$elist 
w_list[1:10] -> MAE@metadata$wlist
WikiMatrix(e_list = MAE@metadata$elist , 
wp_list = MAE@metadata$wlist) -> MAE@ExperimentList$Wmat

TurnPercent(wikiMatrix = MAE@ExperimentList$Wmat,
rowInt = 4) -> MAE@metadata$Pmat

CreateClusters(method = "c", MAE, Percent_matrix = MAE@metadata$Pmat,
no.clusters = 2, Variance = 0.99) -> MAE
}
