% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMiningMatrix.R
\name{DataMiningMatrix}
\alias{DataMiningMatrix}
\title{DataMiningMatrix}
\usage{
DataMiningMatrix(corrTable, targetscan , mirdb, mirtarbase)
}
\arguments{
\item{corrTable}{Correlation matrix between the mRNAs from the
wikipathway of interest and miRNA data.}

\item{targetscan}{Targetscans predicted interactions.}

\item{mirdb}{miRDB predcited interactions.}

\item{mirtarbase}{miRTarBase functional interactions.}
}
\value{
A matrix of information which cross references the occurance of
microRNA-mRNA interactions between databases and the given data.
}
\description{
Mines out predicted/ functional interactions which correspond
between Targetscans, miRDB, miRTarBase and the interactions data frame.
}
