% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpressRes.R
\name{diffExpressRes}
\alias{diffExpressRes}
\title{diffExpressRes}
\usage{
diffExpressRes(MAE, df, dataType = '', genes_ID, idColumn = '', name = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment Object.}

\item{df}{mRNA or miR dataframe (rownames as genes and experimental
variables as columns.)}

\item{dataType}{Column name to take an average from e.g. Log2FC, AveExp,
pval.}

\item{genes_ID}{Dataframe with that was created earlier e.g. mRNA_ensembl
or miR_entrez.}

\item{idColumn}{Name of column to use as the merge point.}

\item{name}{= New name of the dataframe.}
}
\value{
Returns the average of whichever data the user was interested e.g.
Average diffExpressResion or average log2fc.
}
\description{
Get average data for one result type.
}
\examples{
library(biomaRt)
library(MultiAssayExperiment)
miR <- mm_miR[1:100,]
mRNA <- mm_mRNA[1:200,]
MAE <- startObject(miR = miR, mRNA = mRNA)
MAE <- getIdsMirMouse(MAE, assay(MAE, 1))
MAE <- getIdsMrnaMouse(MAE, assay(MAE, 2), "useast")
MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
               genes_ID = assay(MAE, 7),
               idColumn = 'GENENAME',
               name = "mRNA_log2fc")
}
