% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadGMT.R
\name{dloadGMT}
\alias{dloadGMT}
\title{dloadGMT}
\usage{
dloadGMT(MAE, speciesInitials = "")
}
\arguments{
\item{MAE}{MultiAssayExperiment to store downlaoded GMT data in.}

\item{speciesInitials}{Either Hs or Mm, respectively to retreive Homo
sapiens or Mus musculus data.}
}
\value{
3 dataframes. 1) genes-wikipathwayIDs 2) wikipathwayIDs-wikipathway
fullnames 3) combination of 1 and 2. All of which can be sotred in an MAE.
}
\description{
Provides mouse or human wikipathway information.
}
\examples{
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE
dloadGMT(MAE, speciesInitial = "Mm") -> MAE
}
