% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadGmt.R
\name{dloadGmt}
\alias{dloadGmt}
\title{dloadGmt}
\usage{
dloadGmt(MAE, speciesInitials = "")
}
\arguments{
\item{MAE}{MultiAssayExperiment to store downlaoded GMT data in.}

\item{speciesInitials}{Either Hs or Mm, respectively to retreive Homo
sapiens or Mus musculus data.}
}
\value{
3 dataframes. 1) genes-wikipathwayIDs 2) wikipathwayIDs-wikipathway
fullnames 3) combination of 1 and 2. All of which can be sotred in an MAE.
}
\description{
Provides mouse or human wikipathway information.
}
\examples{
miR <- mm_miR
mRNA <- mm_mRNA
MAE <- startObject(miR = miR, mRNA = mRNA)
MAE <- dloadGmt(MAE, speciesInitial = "Mm")
}
