% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadmiRDB.R
\name{dloadmiRDB}
\alias{dloadmiRDB}
\title{dloadmiRDB}
\usage{
dloadmiRDB()
}
\value{
A dataframe of predicted miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (6.0) of predicted targets from
targetscan database.
}
