% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadMirdb.R
\name{dloadMirdb}
\alias{dloadMirdb}
\title{dloadMirdb}
\usage{
dloadMirdb(MAE, species, orgDB)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will have downloaded mirDB data stored
within it. It is recommended to use the MAE which was created using the
mirMrnaInt function.}

\item{species}{Species of interest e.g. "hsa" or "mmu".}

\item{orgDB}{organism db specific for species being looked into. Specific
db must be loaded before use.}
}
\value{
A dataframe of species specific mRNA-miR interactions, and other
information associated with these interactions.
}
\description{
Downloads most recent version (6.0) of predicted targets from
mirdb database http://mirdb.org/download.html. This will take some time.
miR-mRNA interactions from the species of interest will be extracted and
stored in a MAE object as an assay. Species associated org.db package must
be loaded.
}
\examples{
\dontrun{

library(org.Mm.eg.db)

MAE <- MultiAssayExperiment()

MAE <-dloadMirdb(MAE, 'mmu', org.Mm.eg.db)

}
}
