% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadMirdb.R
\name{dloadMirdb}
\alias{dloadMirdb}
\title{dloadMirdb}
\usage{
dloadMirdb(MAE)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}
}
\value{
A dataframe of predicted miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (6.0) of predicted targets from
targetscan database.
}
\examples{
\dontrun{
MAE <- MultiAssayExperiment()
MAE <-dloadMirdb(MAE)
}
}
