% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadMirtarbase.R
\name{dloadMirtarbase}
\alias{dloadMirtarbase}
\title{dloadMirtarbase}
\usage{
dloadMirtarbase(MAE)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}
}
\value{
Dataframe of functionally tested miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (7.0) of functional targets from
miRTarBase database.
}
\examples{
\dontrun{

MAE <- MultiAssayExperiment()

MAE <-dloadMirdb(MAE)

}
}
