% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadMirtarbase.R
\name{dloadMirtarbase}
\alias{dloadMirtarbase}
\title{dloadMirtarbase}
\usage{
dloadMirtarbase(MAE, species)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the downloaded mirtarbase
data. It is recommended to use the MAE which was used in
the mirMrnaInt function.}

\item{species}{Species of interest e.g. "hsa" or "mmu".}
}
\value{
Dataframe of species specific miR-mRNA interactions with
strong functional evidence. Output will be stored as an assay in the input
MAE.
}
\description{
Downloads most recent version (8.0) of functional targets from
the miRTarBase database http://mirtarbase.cuhk.edu.cn/php/download.php
. Species specific miR-mRNA interactions which do not have 'weak' evidence
are used.
}
\examples{
\dontrun{

MAE <- MultiAssayExperiment()

MAE <- dloadMirtarbase(MAE, "mmu")

}
}
