% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadmiRTarBase.R
\name{dloadmiRTarBase}
\alias{dloadmiRTarBase}
\title{dloadmiRTarBase}
\usage{
dloadmiRTarBase()
}
\value{
Dataframe of functionally tested miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (7.0) of functional targets from
miRTarBase database.
}
