% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadTargetscan.R
\name{dloadTargetscan}
\alias{dloadTargetscan}
\title{dloadTargetscan}
\usage{
dloadTargetscan(MAE, species)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the downloaded targetscan
data. It is recommended to use the MAE which was used in
the mirMrnaInt function.}

\item{species}{Species of interest e.g "hsa" or "mmu."}
}
\value{
Dataframe of species specific predicted mRNA-miR interactions. Output
will be stored as an assay in the input MAE.
}
\description{
Downloads most recent version (7.2) of predicted targets from
the targetscan database
http://www.targetscan.org/cgi-bin/targetscan/data_download.vert72.cgi.
miR-mRNA interactions from the species of interest will be extracted.
}
\examples{
\dontrun{

MAE <- MultiAssayExperiment()

MAE <-dloadTargetscan(MAE, "mmu")
}
}
