% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dloadTargetscan.R
\name{dloadTargetscan}
\alias{dloadTargetscan}
\title{dloadTargetscan}
\usage{
dloadTargetscan(MAE)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}
}
\value{
A dataframe of predicted miR-mRNA interactions in different
species.
}
\description{
Downloads most recent version (7.2) of predicted targets from
targetscan database.
}
\examples{
\dontrun{
MAE <- MultiAssayExperiment()
MAE <-dloadTargetscan(MAE)
}
}
