\name{e_list}
\alias{e_list}
\docType{data}
\title{
e_list
}
\description{
List of entrezIDs to make examples run faster.
}
\usage{data("e_list")}
\format{
The format is:
List of 5:
D1
D2
D3
D7
D14
}
\details{
List of entrezgeneIDs per time point in data.
}
\source{
From data.
}
\references{
NA
}
\examples{
data(e_list)
}
\keyword{datasets}
