% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNames.R
\name{eNames}
\alias{eNames}
\title{eNames}
\usage{
eNames(MAE, method = '', gene_IDs, ID_Column)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store output of eNames.
It is recommended to use the MAE which stores the output of addIds.}

\item{method}{Either 'c' or 's', respectively for combined or separated
analysis.}

\item{gene_IDs}{List of DE data and associated gene IDs. This is the
output from the addIds function, this should be found as metadata in
the MAE used in the addIds function.}

\item{ID_Column}{Integer representing the gene ID column in each dataframe.
This will be the last column in each dataframe, but will vary based on the
input data. This should be 2+ the number of DE results per
time point. e.g. if each time point has log2fc's and adjPvalue's, then the
fourth column will contain the gene ID information.}
}
\value{
A single list of entrez/ ensembl IDs for each time point. Output
will be stored as metadata in the input MAE.
}
\description{
Extracts the gene IDs from nested dataframes created from the
addIds function.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

mRNA <- mm_mRNA

Data <- startObject(miR = miR, mRNA = mRNA)

Data <- getIdsMir(Data, assay(Data, 1), org.Mm.eg.db, 'mmu')

Data <- getIdsMrna(Data, assay(Data, 2), mirror = 'useast', 'mmusculus')

Data <- combineGenes(MAE = Data, miR_data = assay(Data, 1),
                     mRNA_data = assay(Data, 2))

Data <- genesList(MAE = Data, method = 'c', genetic_data = assay(Data, 9),
                  timeString = 'D')

Data <- significantVals(MAE = Data, method = 'c',
                        geneList = metadata(Data)[[1]],
                        maxVal = 0.05, stringVal = "adjPVal")

Data <- addIds(MAE = Data, method = "c",
              filtered_genelist = metadata(Data)[[2]],
              miR_IDs = assay(Data, 3), mRNA_IDs = assay(Data, 7))

Data <- eNames(MAE = Data, method = "c", gene_IDs = metadata(Data)[[3]],
               ID_Column = 4)
}
