% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNames.R
\name{eNames}
\alias{eNames}
\title{eNames}
\usage{
eNames(MAE, method = '', gene_IDs, ID_Column)
}
\arguments{
\item{MAE}{MultiAssayExperiment Object.}

\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{gene_IDs}{A nested list of dataframes(c)/ list of lists of nested
dataframes (s), and each dataframe has a column for entrez or ensembl ID.}

\item{ID_Column}{The entrez/ ensembl ID column in each dataframe.
Should be the last column.}
}
\value{
A single list of entrez/ ensembl IDs for each dataframe
within the lists.
}
\description{
Extract the ID names from the nested dataframes within the
metadata of the MAE object.
}
\examples{
library(clusterProfiler)
library(org.Mm.eg.db)
library(MultiAssayExperiment)

miR <- mm_miR
mRNA <- mm_mRNA
Data <- startObject(miR = miR, mRNA = mRNA)

Data <- combineGenes(MAE = Data, miR_data = assay(Data, 1),
                     mRNA_data = assay(Data, 2))

Data <- genesList(MAE = Data, method = 'c', genetic_data = assay(Data, 3),
                  timeString = 'D')

Data <- significantVals(MAE = Data, method = 'c',
                        geneList = metadata(Data)[[1]],
                        maxVal = 0.05, stringVal = "adjPVal")

Data <- addIds(MAE = Data, method = "c",
              filtered_genelist = metadata(Data)[[2]],
              miR_IDs = assay(Data, 3), mRNA_IDs = assay(Data, 3))

Data <- eNames(MAE = Data, method = "c", gene_IDs = metadata(Data)[[3]],
               ID_Column = 4)
}
