% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNames.R
\name{eNames}
\alias{eNames}
\title{eNames}
\usage{
eNames(method = '', gene_IDs, ID_Column)
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{gene_IDs}{A nested list of dataframes(c)/ list of lists of nested
dataframes (s), and each dataframe has a column for entrez or ensembl ID.}

\item{ID_Column}{The entrez/ ensembl ID column in each dataframe.
Should be the last column.}
}
\value{
A single list of entrez/ ensembl IDs for each dataframe
within the lists.
}
\description{
Extract the ID names from the nested dataframes within the
metadata of the MAE object.
}
\examples{
library(biomaRt)
library(clusterProfiler)
library(org.Mm.eg.db)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE
getIDs_miR_mouse(MAE = MAE, miR = MAE@ExperimentList$miR) -> MAE
getIDs_mRNA_mouse(MAE = MAE, mRNA = MAE@ExperimentList$mRNA) -> MAE

CombineGenes(miR_data = MAE@ExperimentList$miR,
mRNA_data = MAE@ExperimentList$mRNA) -> MAE@ExperimentList$genetic_data

GenesList(method = 'c', genetic_data = MAE@ExperimentList$genetic_data,
timeString = 'D') -> MAE@metadata$genelist

SignificantVals(method = "c", geneList = MAE@metadata$genelist,
maxVal = 0.05, stringVal = "adjPVal") -> MAE@metadata$filtered_genelist

AddIDs(method = "c", filtered_genelist = MAE@metadata$filtered_genelist,
miR_IDs = MAE@ExperimentList$miR_entrez,
mRNA_IDs = MAE@ExperimentList$mRNA_entrez) -> MAE@metadata$gene_entrez

eNames(method = "c", gene_IDs = MAE@metadata$gene_entrez, ID_Column = 4
) -> MAE@metadata$e_list
}
