% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichWiki.R
\name{enrichWiki}
\alias{enrichWiki}
\title{enrichWiki}
\usage{
enrichWiki(MAE, method = '', ID_list, orgDB, path_gene, path_name,
                  ID = '', universe, pvalcutoff, qvaluecutoff,
                  padjustmethod)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the output from enrichWiki.
It is recommended to use the MAE object which stores the output from the
dloadGmt function.}

\item{method}{Either 'c' or 's', respectively for combined or separated
analysis.}

\item{ID_list}{List of ensembl or entrez IDs for each sample.
This is the output from eNames function. This will be found as
metadata within the MAE used in the eNames function.}

\item{orgDB}{DB package of the species being analysed. e.g. org.Mm.eg.db
if mouse miR-mRNA data is being looked into.}

\item{path_gene}{Dataframe containing pathway ID - gene ID information. This
is output from either dloadGmt or gmtEnsembl. It will be stored as an assay
within the MAE used in dloadGmt or gmtEnsembl.}

\item{path_name}{Dataframe containing pathway ID - pathway names information.
This is output from dloadGmt. It will be stored as an assay within the MAE
used in dloadGmt.}

\item{ID}{Either "ENTREZID" or "ENSEMBL". This should be the same as the ID
type used for ID_list. dloadGmt loads data as entrez gene IDs and gmtEnsembl
converts this to ensembl gene IDs.}

\item{universe}{A column of gene IDs to be used as the background for
gene set enrichment. IDs should be stored as characters. It is recommended to
use all genes found within the wikipathways of the species being analysed as
background i.e. path_gene$gene or universe = assay(MAE, i)[[2]]/ MAE[[i]][2].
To add a unique universe, create a list of gene IDs (entrezID or ensembl)
which are classed as characters.}

\item{pvalcutoff}{Default is 0.05. P value cut-off point.}

\item{qvaluecutoff}{Default is 0.2. q value cut-off point.}

\item{padjustmethod}{Default is 'BH'. This sets the pvalue adjustment
method. Look into the enricher function from clusterProfiler for more info.}
}
\value{
A large list which identifies which wikipathways are most enriched
at each time point of the input data. Output will be stored as metadata in
the input MAE.
}
\description{
Finds which wikipathways are enriched within the data. This
function uses gene set enrichment analysis from clusterProfiler to
find enriched signalling pathways. Each time point is analysed
individually. In the case of separated TimiRGeN analysis, each gene type and
time point are analysed individually.
}
\examples{
library(org.Mm.eg.db)

MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list_mouse

MAE <- dloadGmt(MAE, species = "Mus musculus")

MAE <- enrichWiki(MAE = MAE, method = 'c', ID_list = metadata(MAE)[[1]],
                  orgDB = org.Mm.eg.db, path_gene = assay(MAE, 1),
                  path_name = assay(MAE, 2), ID = "ENTREZID",
                  universe = assay(MAE, 1)[[2]])
}
