% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_miR_human.R
\name{getIDs_miR_human}
\alias{getIDs_miR_human}
\title{getIDs_miR_human}
\usage{
getIDs_miR_human(MAE, miR)
}
\arguments{
\item{MAE}{MultiAssayObject created by StartObject}

\item{miR}{Dataframe. Rownames are genes and columns are 
results of DE.}
}
\value{
MAE object with 4 more dataframes consisting of ID information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data for human
microRNAs. It will aslo produce adjusted ensembl and entrez for IDs that are
specific to multiple microRNAs.
}
\examples{
library(clusterProfiler)
library(org.Hs.eg.db)
library(MultiAssayExperiment)
miR <- hs_miR
rownames(miR) <- gsub(rownames(miR), pattern = "\\\\.", replacement = "-")
rownames(miR) <- sub("-$", "*", rownames(miR))
miR <- miR[1:100,]
MAE <- StartObject(miR = miR, mRNA = NULL)
MAE <- getIDs_miR_human(MAE, assay(MAE, 1))
}
