% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_miR_human.R
\name{getIDs_miR_human}
\alias{getIDs_miR_human}
\title{getIDs_miR_human}
\usage{
getIDs_miR_human(miR)
}
\arguments{
\item{miR}{Dataframe. Rownames are genes and columns are results of DE.}
}
\value{
4 dataframes consisting of ID information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data for human
microRNAs. It will aslo produce adjusted ensembl and entrez for IDs that are
specific to multiple microRNAs.
}
\examples{
library(biomaRt)
library(org.Hs.eg.db)
library(clusterProfiler)
hs_miR -> miR
gsub(rownames(miR), pattern = "\\\\.", replacement =  "-") -> rownames(miR)
getIDs_miR_human(miR)
}
