% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_miR_mouse.R
\name{getIDs_miR_mouse}
\alias{getIDs_miR_mouse}
\title{getIDs_miR_mouse}
\usage{
getIDs_miR_mouse(MAE, miR)
}
\arguments{
\item{MAE}{MultiAssayObject created by StartObject}

\item{miR}{Dataframe. Rownames are genes and columns are results of DE.}
}
\value{
MAE object with 4 more dataframes consisting of ID information.
}
\description{
getIDs_miR_human will produce ensembl and entrez data for mouse
microRNAs. It will aslo produce adjusted ensembl and entrez for IDs that
are specific to multiple microRNAs.
}
\examples{
library(MultiAssayExperiment)
library(clusterProfiler)
library(org.Mm.eg.db)
miR <- mm_miR
miR <- miR[1:100,]
MAE <- StartObject(miR = miR, mRNA = NULL)
MAE <- getIDs_miR_mouse(MAE, assay(MAE, 1))
}
