% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs_mRNA_mousetohuman.R
\name{getIDs_mRNA_mousetohuman}
\alias{getIDs_mRNA_mousetohuman}
\title{getIDs_mRNA_mousetohuman}
\usage{
getIDs_mRNA_mousetohuman(mRNA, mirror)
}
\arguments{
\item{mRNA}{Dataframe. Rownames should be genenames and colnames should be
results from DE.}

\item{mirror}{String to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
3 dataframes. One with human entrez information, another with human
ensembl information and a new dataframe of human named genes with the DE
results.
}
\description{
getIDs_miR_mousetohuman will produce ensembl and entrez data
for mouse mRNAs. Also a new dataframe of human named mRNAs will be made.
}
\examples{
library(biomaRt)
mm_mRNA -> mRNA
mRNA[1:200,] -> mRNA
getIDs_mRNA_mousetohuman(mRNA = mRNA, mirror = 'useast')
}
