% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMirMouse.R
\name{getIdsMirMouse}
\alias{getIdsMirMouse}
\title{getIdsMirMouse}
\usage{
getIdsMirMouse(MAE, miR)
}
\arguments{
\item{MAE}{MultiAssayExperiment to store the output of getIdsMirMouse.
It is recommended to use the MAE which contains output from startObject.}

\item{miR}{A Dataframe. Rownames are genes and columns are results of DE.
This should be found as an assay within the MAE used in the startObject
function. Please read vignette for nomenclature guidance.}
}
\value{
4 dataframes consisting of either entrez or ensembl ID information.
2 of these will be adjusted for shared IDs. Output will be stored as assays
in the input MAE.
}
\description{
getIdsMirMouse will produce ensembl and entrez ID data for
mouse microRNAs. It will also produce adjusted ensembl and entrez for IDs
that are specific to microRNAs that share an ID. They will be stored as
4 individual assays in a MAE. org.Mm.eg.db must be loaded prior to using
this function.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

# Make sure miRNA gene name nomenclature is correct for TimiRGeN analysis!

miR <- miR[1:100,]

MAE <- startObject(miR = miR, mRNA = NULL)

MAE <- getIdsMirMouse(MAE, assay(MAE, 1))
}
