% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsMrnaHuman.R
\name{getIdsMrnaHuman}
\alias{getIdsMrnaHuman}
\title{getIdsMrnaHuman}
\usage{
getIdsMrnaHuman(MAE, mRNA, mirror)
}
\arguments{
\item{MAE}{MultiAssayExperiment to store the output of getIdsMrnaHuman.
It is recommended to use the MAE which contains output from startObject.}

\item{mRNA}{A Dataframe. Rownames are genes and columns are results of DE.
This should be found as an assay within the MAE used in the startObject
function. Please read vignette for nomenclature guidance.}

\item{mirror}{String to identify which biomaRt repo is best for the users
locations. Either 'useast', 'uswest', 'asia' or 'www'. Default is 'useast'.}
}
\value{
2 new dataframes. One with entrez information and another with
ensembl information.
}
\description{
getIdsMrnaHuman will produce ensembl and entrez data for
human mRNAs.These will be stored as 2 individual assays within a MAE.
}
\examples{
mRNA <- mm_mRNA

mRNA <- mRNA[1:20,]

MAE <- startObject(miR = NULL, mRNA = mRNA)

MAE <- getIdsMrnaHuman(MAE = MAE, mRNA = assay(MAE, 2), mirror = 'www')
}
