% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT_ensembl.R
\name{GMT_ensembl}
\alias{GMT_ensembl}
\title{GMT_ensembl}
\usage{
GMT_ensembl(path_gene, path_data, orgDB)
}
\arguments{
\item{path_gene}{wikipathway-gene data from downloadGMT.}

\item{path_data}{wikipathway-gene-wikipathway full name data from
downloadGMT.}

\item{orgDB}{organism Library e.g. org.Hs.eg.db}
}
\value{
2 dataframes. Wikipathway data based on ensembl IDs.
}
\description{
Change entrez IDs in wpid2gene and wp_data into ensembl IDs
}
\examples{
library(org.Mm.eg.db)
downloadGMT(speciesInitial = "Mm")
GMT_ensembl(path_gene, path_data, org.Mm.eg.db)
file.remove("mus.gmt")
}
