% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtEnsembl.R
\name{gmtEnsembl}
\alias{gmtEnsembl}
\title{gmtEnsembl}
\usage{
gmtEnsembl(MAE, path_gene, path_data, orgDB)
}
\arguments{
\item{MAE}{MultiAssayExperiment where the ensembl wikipathways data can
be stored.}

\item{path_gene}{path_gene data from downloadGMT.}

\item{path_data}{path_data from downloadGMT.}

\item{orgDB}{organism Library e.g. org.Hs.eg.db}
}
\value{
2 dataframes. Wikipathway data based on ensembl IDs.
}
\description{
Change entrez IDs in path_gene and path_data into ensembl IDs.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

mRNA <- mm_mRNA

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- dloadGmt(MAE, speciesInitial = "Mm")

MAE <- gmtEnsembl(MAE = MAE, assay(MAE, 3),
                   assay(MAE, 5), org.Mm.eg.db)
}
