% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtEnsembl.R
\name{gmtEnsembl}
\alias{gmtEnsembl}
\title{gmtEnsembl}
\usage{
gmtEnsembl(MAE, path_gene, path_data, orgDB)
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the output of gmtEnsembl.
It is recommended to use the same MAE object which contains output from
dloadGmt.}

\item{path_gene}{Dataframe with wikipathway IDs and entrezgene IDs.
path_gene is from the dloadGmt function. It will be stored as an assay
within the MAE used in the dloadGmt function.}

\item{path_data}{Dataframe with wikipathway IDs, wikipathway names and
entrezgene IDs. path_data is from the dloadGmt function. It will be stored
as an assay within the MAE used in the dloadGmt function.}

\item{orgDB}{Load the appropriate db package e.g. org.Hs.eg.db if human
wikipathways are being used.}
}
\value{
2 dataframes. One containing wikipathway IDs and ensembl gene IDs,
and the other containing  wikipathway IDs, ensembl gene IDs and wikipathway
names. Output will be stored as assays in the input MAE.
}
\description{
Change entrez IDs in path_gene and path_data into ensembl IDs.
Will create two new dataframes with ensembl IDs and wikipathway information.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

mRNA <- mm_mRNA

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- dloadGmt(MAE, species = "Mus musculus")

MAE <- gmtEnsembl(MAE = MAE, assay(MAE, 3),
                   assay(MAE, 5), org.Mm.eg.db)
}
