% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDynamic.R
\name{MakeDynamic}
\alias{MakeDynamic}
\title{MakeDynamic}
\usage{
MakeDynamic(miR_expression, mRNA_expression, miR_IDs_adj,
Datatype = '')
}
\arguments{
\item{miR_expression}{microRNA data after going through Express function.}

\item{mRNA_expression}{mRNA data after going through Express function.}

\item{miR_IDs_adj}{Either miR_ensembl_adj or miR_entrez_adj}

\item{Datatype}{Either En (ensembl data) or L (entrez data)}
}
\value{
MicroRNA and mRNA dynamic data that can be used in pathvisio if
exported.
}
\description{
Make expression data input for pathvisio so changes over time
can be visualised.
}
\examples{
library(org.Mm.eg.db)
library(clusterProfiler)
library(biomaRt)

mm_miR -> miR
mm_mRNA -> mRNA

StartObject(miR = miR, mRNA = mRNA) -> MAE

getIDs_mRNA_mouse(MAE, MAE@ExperimentList$mRNA, "useast") -> MAE
getIDs_miR_mouse(MAE, MAE@ExperimentList$miR) -> MAE

Express(df = MAE@ExperimentList$mRNA, dataType = 'Log2FC',
genes_ID = MAE@ExperimentList$mRNA_entrez,
idColumn = 'GENENAME') -> MAE@ExperimentList$mRNA_express
Express(df = MAE@ExperimentList$miR, dataType = 'Log2FC',
genes_ID = MAE@ExperimentList$miR_entrez,
idColumn = 'GENENAME') -> MAE@ExperimentList$miR_express

MakeDynamic(miR_expression = MAE@ExperimentList$miR_express, 
mRNA_expression = MAE@ExperimentList$mRNA_express, 
miR_IDs_adj = MAE@ExperimentList$miR_entrez, Datatype = "L"
) -> MAE@ExperimentList$Dynamics
}
