% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDynamic.R
\name{MakeDynamic}
\alias{MakeDynamic}
\title{MakeDynamic}
\usage{
MakeDynamic(miR_expression, mRNA_expression, miR_IDs_adj,
Datatype = '')
}
\arguments{
\item{miR_expression}{microRNA data after going through Express function.}

\item{mRNA_expression}{mRNA data after going through Express function.}

\item{miR_IDs_adj}{Either miR_ensembl_adj or miR_entrez_adj}

\item{Datatype}{Either En (ensembl data) or L (entrez data)}
}
\value{
MicroRNA and mRNA dynamic data that can be used in pathvisio if
exported.
}
\description{
Make expression data input for pathvisio so changes over time
can be visualised.
}
\examples{
library(biomaRt)
library(org.Mm.eg.db)
library(clusterProfiler)
mm_mRNA -> mRNA
mm_miR -> miR
getIDs_mRNA_mouse(mRNA)
getIDs_miR_mouse(miR)
Express(df = mRNA, dataType = 'Log2FC', genes_ID = mRNA_entrez,
idColumn = 'GENENAME') -> mRNA_express
Express(df = miR, dataType = 'Log2FC', genes_ID = miR_entrez,
idColumn = 'GENENAME') -> miR_express
MakeDynamic(miR_expression = miR_express,
mRNA_expression = mRNA_express,
miR_IDs = miR_adjusted_entrez,
Datatype = 'L') -> Dynamics
}
