% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDynamic.R
\name{makeDynamic}
\alias{makeDynamic}
\title{makeDynamic}
\usage{
makeDynamic(MAE, miR_expression, mRNA_expression, miR_IDs_adj,
                   dataType = '')
}
\arguments{
\item{MAE}{MultiAssayObject.}

\item{miR_expression}{microRNA data after going through Express function.}

\item{mRNA_expression}{mRNA data after going through Express function.}

\item{miR_IDs_adj}{Either miR_ensembl_adj or miR_entrez_adj}

\item{dataType}{Either En (ensembl data) or L (entrez data)}
}
\value{
MicroRNA and mRNA dynamic data that can be used in pathvisio if
exported.
}
\description{
Make expression data input for pathvisio so changes over time
can be visualised.
}
\examples{
miR <- mm_miR[1:100,]
mRNA <- mm_mRNA[1:200,]
MAE <- startObject(miR = miR, mRNA = mRNA)
MAE <- getIdsMirMouse(MAE, assay(MAE, 1))
MAE <- getIdsMrnaMouse(MAE, assay(MAE, 2), "useast")
MAE <- diffExpressRes(MAE, df = assay(MAE, 1), dataType = 'Log2FC',
               genes_ID = assay(MAE, 3),
               idColumn = 'GENENAME',
               name = "miR_express")
MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
               genes_ID = assay(MAE, 7),
               idColumn = 'GENENAME',
               name = 'mRNA_express')

MAE <- makeDynamic(MAE, miR_expression = assay(MAE, 9),
                  mRNA_expression = assay(MAE, 10),
                  miR_IDs_adj = assay(MAE, 5),
                  dataType = "L")
}
