% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDynamic.R
\name{makeDynamic}
\alias{makeDynamic}
\title{makeDynamic}
\usage{
makeDynamic(MAE, miR_expression, mRNA_expression, miR_IDs_adj,
                   dataType = '')
}
\arguments{
\item{MAE}{MultiAssayExpreriment to store the output of makeDynamic. It is
recommended to use the same MAE which stores output from matrixFilter.}

\item{miR_expression}{Dataframe containing abundance values
(e.g. log2fc or average expression) from miR specific differential expression
, along with gene IDs. This is the output from diffExpressRes. Output of
diffExpressRes should be stored as an assay within the MAE used in
diffExpressRes.}

\item{mRNA_expression}{Dataframe containing abundance values (log2fc or
average expression) from mRNA specific differential expression, along with
gene IDs. This is the output from diffExpressRes. Output of diffExpressRes
should be stored as an assay within the MAE used in diffExpressRes.}

\item{miR_IDs_adj}{Dataframe which contain adjusted gene IDs from miR data.
Either miR_adjusted_entrez or miR_adjusted_ensembl. Should be found
as an assay in the MAE used a getIdsMir function.}

\item{dataType}{String which represents the gene ID used in this analysis.
Either "En" (ensembl data) or "L" (entrez data).}
}
\value{
miR and mRNA dynamic data that can be saved and be used in
pathvisio to display dynamic behaviour of miRs and mRNAs of interest over
the time series in a signalling pathway of interest. Output will be stored
as an assay in the input MAE.
}
\description{
Produces a dataframe that can be imported into pathvisio
to show how changes in genes expression levels over the time course. Follow
instructions found in the vignette which show on how to save this file and
further instructions found in /inst/Pathvisio_GRN_guide.pdf to see how
this can help in GRN construction.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR[1:100,]

mRNA <- mm_mRNA[1:200,]

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- getIdsMirMouse(MAE, assay(MAE, 1))

MAE <- getIdsMrnaMouse(MAE, assay(MAE, 2), "useast")

MAE <- diffExpressRes(MAE, df = assay(MAE, 1), dataType = 'Log2FC',
               genes_ID = assay(MAE, 3),
               idColumn = 'GENENAME',
               name = "miR_express")

MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
               genes_ID = assay(MAE, 7),
               idColumn = 'GENENAME',
               name = 'mRNA_express')

MAE <- makeDynamic(MAE, miR_expression = assay(MAE, 9),
                  mRNA_expression = assay(MAE, 10),
                  miR_IDs_adj = assay(MAE, 5),
                  dataType = "L")
}
