% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMapp.R
\name{makeMapp}
\alias{makeMapp}
\title{makeMapp}
\usage{
makeMapp(MAE, filt_df, miR_IDs_adj, dataType = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment to store the output of makeMapp. It
is recommended to use the same MAE which stores the output from
matrixFilter.}

\item{filt_df}{Dataframe of mined miR-mRNA interactions. This is output
of matrixFilter. It should be stored as an assay in the MAE used in the
matrixFilter function.}

\item{miR_IDs_adj}{Dataframes with adjusted gene IDs to account for -5p and
-3p specific miRs. miR_adjusted_entrez or miR_adjusted_ensembl. Should be
found as assays within the MAE used a getIdsMir function.}

\item{dataType}{String which represents the gene ID used in this analysis.
Either "En" (ensembl data) or "L" (entrez data).}
}
\value{
A dataframe containing microRNAs and adjusted gene IDs which can be
saved as a text file to be imported into pathvisio via the MAPPapp. Output
will be saved as an assay in the input MAE.
}
\description{
Creates a dataframe which can be imported into pathvisio by use
of the the MAPP plugin. This will add the filtered miRs to the wikipathway of
interest on pathvisio. Follow instructions found in the vignette which show
how to save this file and further instructions found in the issues section
of the TimiRGeN gihub https://github.com/Krutik6/TimiRGeN/issues/2.
}
\examples{
library(org.Mm.eg.db)

data(mm_miR)

data(mm_mRNA)

MAE <- startObject(miR = mm_miR, mRNA = mm_mRNA)

MAE <- getIdsMir(MAE, assay(MAE, 1), orgDB = org.Mm.eg.db, 'mmu')

MAE <- getIdsMrna(MAE, assay(MAE, 2), "useast", 'mmusculus')

Filt_df <- data.frame(row.names = c("mmu-miR-320-3p:Acss1",
                                     "mmu-miR-27a-3p:Odc1"),
                      avecor = c(-0.9191653, 0.7826041),
                      miR = c("mmu-miR-320-3p", "mmu-miR-27a-3p"),
                      mRNA = c("Acss1", "Odc1"),
                      miR_Entrez = c(NA, NA),
                      mRNA_Entrez = c(68738, 18263),
                      TargetScan = c(1, 0),
                      miRDB = c(0, 0),
                      Predicted_Interactions = c(1, 0),
                      miRTarBase = c(0, 1),
                      Pred_Fun = c(1, 1))

MAE <- makeMapp(MAE, filt_df = Filt_df, miR_IDs_adj = assay(MAE, 5),
                dataType = 'L')
}
