% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMapp.R
\name{makeMapp}
\alias{makeMapp}
\title{makeMapp}
\usage{
makeMapp(MAE, filt_df, miR_IDs_adj, dataType = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment object to store the output of makeMapp. It
is recommended to use the same MAE object which stores the results from
matrixFilter.}

\item{filt_df}{Dataframe of mined microRNA-mRNA interactions. This is output
of matrixFilter. It should be stored as an assay in the MAE used in the
matrixFilter function.}

\item{miR_IDs_adj}{Dataframes with adjusted gene IDs to account for -5p and
-3p specific miRs. miR_adjusted_entrez or miR_adjusted_ensembl. Should be
found as an assay in the MAE used a getIdsMir function.}

\item{dataType}{String which represents the gene ID used in this analysis.
Either "En" (ensembl data) or "L" (entrez data).}
}
\value{
A dataframe containing microRNA and adjusted gene ID information
which should be saved as a text file for import into pathvisio via the MAPP
app.
}
\description{
Creates input for the MAPP plugin in pathvisio. This will
help to add the filtered miRNAs to the wikipathway of interest.
Follow vignette instructions on how to save this file and further
instructions found in /inst/Pathvisio_GRN_guide.pdf to see how
this can help in GRN construction.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR

mRNA <- mm_mRNA

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- getIdsMirMouse(MAE, assay(MAE, 1))

Filt_df <- data.frame(row.names = c("mmu-miR-320-3p:Acss1",
                                     "mmu-miR-27a-3p:Odc1"),
                      avecor = c(-0.9191653, 0.7826041),
                      miR = c("mmu-miR-320-3p", "mmu-miR-27a-3p"),
                      mRNA = c("Acss1", "Acss1"),
                      miR_Entrez = c(NA, NA),
                      mRNA_Entrez = c(68738, 18263),
                      TargetScan = c(1, 0),
                      miRDB = c(0, 0),
                      Predicted_Interactions = c(1, 0),
                      miRTarBase = c(0, 1),
                      Pred_Fun = c(1, 1))

MAE <- makeMapp(MAE, filt_df = Filt_df, miR_IDs_adj = assay(MAE, 5),
                dataType = 'L')
}
