% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNet.R
\name{makeNet}
\alias{makeNet}
\title{makeNet}
\usage{
makeNet(MAE, filt_df)
}
\arguments{
\item{MAE}{MultiAssayExperiment Object.}

\item{filt_df}{Filtered miR-mRNA interactions from the mining matrix
produced by the DataMiningMatrix function.}
}
\value{
A list which is the input for Quicknet
}
\description{
Creates an igraph object from filtered miR-mRNA interactions
from database mining.
}
\examples{
library(igraph)
library(biomaRt)
library(MultiAssayExperiment)

Filt_df <- data.frame(row.names = c("mmu-miR-320-3p:Acss1",
                                     "mmu-miR-27a-3p:Odc1"),
                      avecor = c(-0.9191653, 0.7826041),
                      miR = c("mmu-miR-320-3p", "mmu-miR-27a-3p"),
                      mRNA = c("Acss1", "Acss1"),
                      miR_Entrez = c(NA, NA),
                      mRNA_Entrez = c(68738, 18263),
                      TargetScan = c(1, 0),
                      miRDB = c(0, 0),
                      Predicted_Interactions = c(1, 0),
                      miRTarBase = c(0, 1),
                      Pred_Fun = c(1, 1))
MAE <- MultiAssayExperiment()
MAE <- makeNet(MAE, Filt_df)
}
