\name{mfuzzdata}
\alias{mfuzzdata}
\docType{data}
\title{
mfuzzdata
}
\description{
mfuzzdata from CreateClusters function in TimiRGeN. For use in passing
vignette build process on bioconductor.
}
\usage{data("mfuzzdata")}
\format{
Expressionset object.
}
\examples{
data(mfuzzdata)
## maybe str(mfuzzdata) ; plot(mfuzzdata) ...
}
\keyword{datasets}
