% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickBar.R
\name{quickBar}
\alias{quickBar}
\title{quickBar}
\usage{
quickBar(X, Y)
}
\arguments{
\item{X}{Dataframes within a list including count information, confidence
scores and wikipathway information. This is the output from the enrichWiki
function. It will be stored as metadata within the MAE used in the enrichWiki
function. Data can be retrieved using [[i]][[j]] on the output of enrichWiki.}

\item{Y}{String which is associated to the nested dataframe selected for X.
This is the output from the enrichWiki function. It will be stored as
metadata within the MAE used in the enrichWiki function. Data can
be retrieved using names([[i]][j]) on the output of enrichWiki.}
}
\value{
Bar plot showing which pathways are most enriched for genes found
at each time point ("c") or at each time point within a gentype ("s").
}
\description{
Creates a bar plot which compares the confidence levels for
each wikipathway association to the filtered input data. The number of genes
in common between a pathway and the input data are taken into account to
generate confidence scores. This function is used specifically for a single
time point at a time, or if using "s" analysis, the function is used for a
single time point within a single gene type (miR or mRNA).
}
\examples{
library(org.Mm.eg.db)

MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list_mouse

MAE <- dloadGmt(MAE, species = "Mus musculus")

MAE <- enrichWiki(MAE = MAE, method = 'c', ID_list = metadata(MAE)[[1]],
                   orgDB = org.Mm.eg.db, path_gene = assay(MAE, 1),
                   path_name = assay(MAE, 2), ID = "ENTREZID",
                   universe = assay(MAE, 1)[[2]])

q <- quickBar(X = metadata(MAE)[[2]][[1]], Y = names(metadata(MAE)[[2]][1]))

# to view bar plot enter plot(q)
}
