% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickBar.R
\name{quickBar}
\alias{quickBar}
\title{quickBar}
\usage{
quickBar(X, Y)
}
\arguments{
\item{X}{Dataframe including count information and wikipathways.}

\item{Y}{Point to the associated name within a list of nested dataframes.}
}
\value{
barplot
}
\description{
Creates dotplot which compares the confidence level we have for
each wikipathways association to the data, the number of genes associated
to each wikipathway and this is all in a ranked fashion for each
timepoint and genetype.
}
\examples{
library(ggplot2)
library(org.Mm.eg.db)
library(clusterProfiler)
library(MultiAssayExperiment)
miR <- mm_miR
mRNA <- mm_mRNA
MAE <- startObject(miR = miR, mRNA = mRNA)

metadata(MAE)[["e_list"]] <- e_list
MAE <- dloadGmt(MAE, speciesInitial = "Mm")

MAE <- enrichWiki(MAE = MAE, method = 'c', ID_list = metadata(MAE)[[1]],
                   orgDB = org.Mm.eg.db, path_gene = assay(MAE, 3),
                   path_name = assay(MAE, 4), ID = "ENTREZID",
                   universe = assay(MAE, 3)[[2]])

q <- quickBar(X = metadata(MAE)[[2]][[1]], Y = metadata(MAE)[[2]][1])
}
