% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quickdot.R
\name{Quickdot}
\alias{Quickdot}
\title{Quickdot}
\usage{
Quickdot(X, Y)
}
\arguments{
\item{X}{Dataframe including count information and wikipathways}

\item{Y}{Point to the associated name within a list of nested dataframes.}
}
\value{
dotplot
}
\description{
Creates dotplot which compares the confidence level we have for
each wikipathways association to the data, the number of genes associated
to each wikipathway and this is all in a ranked fashion for each timepoint
and genetype.
}
\examples{
library(ggplot2)
library(org.Mm.eg.db)
library(clusterProfiler)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE

e_list -> MAE@metadata$elist
dloadGMT(MAE, speciesInitial = "Mm") -> MAE

MAE@metadata$sigwiki <- EnrichWiki(method = "c",
e_list = MAE@metadata$elist,
orgDB = org.Mm.eg.db, 
path_gene = MAE@ExperimentList$path_gene, 
path_name = MAE@ExperimentList$path_name, 
ID = "ENTREZID", 
universe = MAE@ExperimentList$path_gene$gene)

Quickdot(X = MAE@metadata$sigwiki$D7_wikipathways, 
Y = MAE@metadata$sigwiki[4]) -> Q
}
