% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quickfuzz.R
\name{Quickfuzz}
\alias{Quickfuzz}
\title{Quickfuzz}
\usage{
Quickfuzz(Mfuzzdata, Clusters, W)
}
\arguments{
\item{Mfuzzdata}{Output after CreateClusters. An Expressionset object.}

\item{Clusters}{Output after CreateClusters. A list.}

\item{W}{Should the plot be shown in a new window? Default is TRUE.}
}
\value{
A plot of different clusters showing how the input data varies
across different wikipathways.
}
\description{
Plots Clusters in reference to the ClusterData dataframe. To be
used after CreateClusters and ClusterCheck. From here individual
wikipathways that show variance throughout a signalling network can be
taken for further miR-mRNA interaction analysis.
}
\examples{
library(Mfuzz)
library(MultiAssayExperiment)
MAE <- Clusters
Quickfuzz(Mfuzzdata = experiments(MAE)[[2]],
          Clusters = metadata(MAE)[[1]], W = FALSE)
}
