% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickFuzz.R
\name{quickFuzz}
\alias{quickFuzz}
\title{quickFuzz}
\usage{
quickFuzz(Mfuzzdata, Clusters, W, background, labelcol, axiscol,
axisline, subcol)
}
\arguments{
\item{Mfuzzdata}{A large ExpressionSet object which contain fuzzy clustering
data. This is output from the createClusters function. The Expressionset
object should be stored as an experiment in the MAE used in the
createClusters function.}

\item{Clusters}{A large list containing information about clusters,
statistics and phenodata. This is output from the createClusters function.
The list should be stored as metadata in the MAE used in the
createClusters function.}

\item{W}{TRUE or FALSE? Should the plot be shown in a new window? Default is
TRUE.}

\item{background}{Plot background colour. Default is black.}

\item{labelcol}{Plot labels colour. Default is yellow.}

\item{axiscol}{Plot axis labels colour. Default is white.}

\item{axisline}{Plot axis line colour. Default is white.}

\item{subcol}{Plot sub title colour. Default is yellow.}
}
\value{
A plot of different clusters showing how the number of genes found
to be significant varies between the input data and wikipathways. These
variations are captured as temporal behaviours and are clustered.
}
\description{
Plots fuzzy clusters. Each different cluster created will
represent a different temporal behaviour. Depending on the data, more or
fewer cluster may be appropriate.
Use clusterCheck to influence this decision before moving onto quickFuzz.
Each line in a cluster represents a pathway. Pathways are divided by colour.
The more intense the colour of a line, the stronger they fit a particular
cluster / temporal behaviour.
Fuzzy clustering is a soft clustering approach where objects are not
divided into fixed clusters. Each pathway can exist in each cluster but
each pathway will differ on the degree to which they fit to each cluster.
Look into the clusterData dataframe created by createClusters to see this.
If a cluster peaks interest, continue to analysis of that cluster with the
returnCluster function.
}
\examples{
MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list_mouse

metadata(MAE)[["w_list"]] <- w_list_mouse[1:10]

MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE,
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)

MAE <- createClusters(MAE, method = "c",
                      percentMatrix = assay(MAE, 2),
                      noClusters = 2, variance = 0.99)

quickFuzz(Mfuzzdata = experiments(MAE)[[4]],
          Clusters = metadata(MAE)[[3]], W = FALSE)
}
