% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickFuzz.R
\name{quickFuzz}
\alias{quickFuzz}
\title{quickFuzz}
\usage{
quickFuzz(Mfuzzdata, Clusters, W)
}
\arguments{
\item{Mfuzzdata}{Output after CreateClusters. An Expressionset object.}

\item{Clusters}{Output after CreateClusters. A list.}

\item{W}{Should the plot be shown in a new window? Default is TRUE.}
}
\value{
A plot of different clusters showing how the input data varies
across different wikipathways.
}
\description{
Plots Clusters in reference to the ClusterData dataframe. To be
used after CreateClusters and ClusterCheck. From here individual
wikipathways that show variance throughout a signalling network can be
taken for further miR-mRNA interaction analysis.
}
\examples{
library(MultiAssayExperiment)
library(Mfuzz)
MAE <- MultiAssayExperiment()
metadata(MAE)[["e_list"]] <- e_list
metadata(MAE)[["w_list"]] <- w_list[1:10]
MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE,
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)

MAE <- createClusters(MAE, method = "c",
                      percentMatrix = assay(MAE, 2),
                      noClusters = 2, variance = 0.99)

quickFuzz(Mfuzzdata = experiments(MAE)[[4]],
          Clusters = metadata(MAE)[[3]], W = FALSE)
}
