% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickNet.R
\name{quickNet}
\alias{quickNet}
\title{quickNet}
\usage{
quickNet(net)
}
\arguments{
\item{net}{network generated by makeNet function. If you have fewer than two
interactions the plot will not work.}
}
\value{
A network depicting miR-mRNA interactions.
}
\description{
Generates an internal R network representing the miR-mRNA
interactions filtered through the input data using wikipathways analysis and
database filtering.
}
\examples{
Filt_df <- data.frame(row.names = c("mmu-miR-320-3p:Acss1",
                                     "mmu-miR-27a-3p:Odc1"),
                      avecor = c(-0.9191653, 0.7826041),
                      miR = c("mmu-miR-320-3p", "mmu-miR-27a-3p"),
                      mRNA = c("Acss1", "Acss1"),
                      miR_Entrez = c(NA, NA),
                      mRNA_Entrez = c(68738, 18263),
                      TargetScan = c(1, 0),
                      miRDB = c(0, 0),
                      Predicted_Interactions = c(1, 0),
                      miRTarBase = c(0, 1),
                      Pred_Fun = c(1, 1))

MAE <- MultiAssayExperiment()

MAE <- makeNet(MAE, Filt_df)

Q <- quickNet(metadata(MAE)[[1]])
}
