% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickNet.R
\name{quickNet}
\alias{quickNet}
\title{quickNet}
\usage{
quickNet(net, pathwayname)
}
\arguments{
\item{net}{List generated by makeNet function. If you have fewer than two
interactions the plot will not work. Output from makeNet should be stored
as metadata within the MAE used in the makeNet function.}

\item{pathwayname}{Character which is the name of pathway of interest.
Default is "Pathway".}
}
\value{
A network depicting filtered miR-mRNA interactions for a specific
wikipathway of interest.
}
\description{
Generates an igraph network representing the miR-mRNA
interactions filtered from the input data using pathway analysis and
database filtering. Pink nodes are miRs and light blue nodes are mRNAs.
Edges are coloured by the average correlations. Note, plot window size may
need to be adjusted to accommodate image.
}
\examples{
Filt_df <- data.frame(row.names = c("mmu-miR-320-3p:Acss1",
                                     "mmu-miR-27a-3p:Odc1"),
                      corr = c(-0.9191653, 0.7826041),
                      miR = c("mmu-miR-320-3p", "mmu-miR-27a-3p"),
                      mRNA = c("Acss1", "Acss1"),
                      miR_Entrez = c(NA, NA),
                      mRNA_Entrez = c(68738, 18263),
                      TargetScan = c(1, 0),
                      miRDB = c(0, 0),
                      Predicted_Interactions = c(1, 0),
                      miRTarBase = c(0, 1),
                      Pred_Fun = c(1, 1))

MAE <- MultiAssayExperiment()

MAE <- makeNet(MAE, Filt_df)

quickNet(metadata(MAE)[[1]])
}
