% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceWiki.R
\name{reduceWiki}
\alias{reduceWiki}
\title{reduceWiki}
\usage{
reduceWiki(MAE, path_data, stringWiki = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{path_data}{File with wikipathway - gene data.}

\item{stringWiki}{Full name of the wikipathway of interest.}
}
\value{
A dataframe which only contains information about the wikipathway of
interest.
}
\description{
Return gene names for a single wikipathway of interest.
}
\examples{
MAE <- MultiAssayExperiment(list(path_data = data.frame(
                        "wpid" = c(rep("WP571", 6)),
                        "gene" = c(16175, 12370,26419, 19249, 19645, 18479),
                        "name" = c(rep("Fas pathway and Stress induction of HSP regulation",
                        6)))))
MAE <- reduceWiki(MAE, path_data = assay(MAE, 1),
                 stringWiki = 'Fas pathway and Stress induction of HSP regulation')
}
