% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnCluster.R
\name{returnCluster}
\alias{returnCluster}
\title{returnCluster}
\usage{
returnCluster(MAE, clusterData, whichCluster, fitCluster)
}
\arguments{
\item{MAE}{MultiAssayExperiment. Results from returnCluster will
be stored within this MAE. It is recommended to use the same MAE which
stores output from createClusters.}

\item{clusterData}{A dataframe which contains cluster-pathway fit scores
and is stored as an assay within the MAE used in the createClusters function.}

\item{whichCluster}{Integer should corresponds to the cluster of interest.}

\item{fitCluster}{How well should the pathways fit into the this
cluster? Integer from 0-1. Default is 0.99.}
}
\value{
A dataframe of which pathways corresponded best with the chosen
dynamics seen in the selected cluster.
}
\description{
Will retrieve information about which wikipathways fitted best
with a specific cluster and store this as an assay within a MAE.
}
\examples{
MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list

metadata(MAE)[["w_list"]] <- w_list_mouse[1:10]

MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE,
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)

MAE <- createClusters(MAE, method = "c",
                    percentMatrix = assay(MAE, 2),
                    noClusters = 2, variance = 0.99)

MAE <- returnCluster(MAE, clusterData = assay(MAE, 3), whichCluster = 1,
                     fitCluster = 0.5)
}
