% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnCluster.R
\name{ReturnCluster}
\alias{ReturnCluster}
\title{ReturnCluster}
\usage{
ReturnCluster(MAE, clusterData, whichCluster, fitCluster)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{clusterData}{Output after CreateClusters. A list.}

\item{whichCluster}{Integer should correspond to the order of clusters
displayed.}

\item{fitCluster}{How well should the wikipathways fit into the this
cluster? Integer from 0-1. Default is 0.99.}
}
\value{
A dataframe of which pathways corresponded best with the chosen
dynamics seen in the selected cluster.
}
\description{
Will retrive information about which wikipathways fitted best
with a specific cluster.
}
\examples{
library(MultiAssayExperiment)
MAE <- Clusters
MAE <- ReturnCluster(MAE, clusterData = assay(MAE, 1), whichCluster = 1, 
                     fitCluster = 0.5)
}
