% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlots.R
\name{savePlots}
\alias{savePlots}
\title{savePlots}
\usage{
savePlots(largeList, maxInt, quickType, fileType = '')
}
\arguments{
\item{largeList}{A list of wikipathways associated to entrezIDs for each
sample.}

\item{maxInt}{Number of samples in dataset/ lists in sigwiki}

\item{quickType}{Quickplot ot Quickbar}

\item{fileType}{Type of file for images to be exported as: png, tiff,
svg or jpg}
}
\value{
saves tiff files of all plots in current dir
}
\description{
Saves all possible plots in the current directory.
}
\examples{
\donttest{
library(org.Mm.eg.db)
miR <- mm_miR
mRNA <- mm_mRNA
MAE <- startObject(miR = miR, mRNA = mRNA) -> MAE

metadata(MAE)[["ID_list"]] <- e_list
MAE <- dloadGmt(MAE, speciesInitial = "Mm")

metadata(MAE)[["sigwiki"]] <- enrichWiki(MAE, method = "c",
                                         ID_list = metadata(MAE)[[1]],
                                         orgDB = org.Mm.eg.db,
                                         path_gene = assay(MAE, 3),
                                         path_name = assay(MAE, 4),
                                         ID = "ENTREZID",
                                         universe = assay(MAE, 3)[2])

savePlots(largeList = metadat(MAE)[[2]], maxInt = 5,
         quickType = quickDot, fileType = "jpg")
}
}
