% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavePlots.R
\name{SavePlots}
\alias{SavePlots}
\title{SavePlots}
\usage{
SavePlots(largeList, maxInt, quickType, fileType = '')
}
\arguments{
\item{largeList}{A list of wikipathways associated to entrezIDs for each
sample.}

\item{maxInt}{Number of samples in dataset/ lists in sigwiki}

\item{quickType}{Quickplot ot Quickbar}

\item{fileType}{Type of file for images to be exported as: png, tiff,
svg or jpg}
}
\value{
saves tiff files of all plots in current dir
}
\description{
Saves all possible plots in the current directory.
}
\examples{
library(biomaRt)
library(ggplot2)
library(org.Mm.eg.db)
library(clusterProfiler)
mm_miR -> miR
mm_mRNA -> mRNA
StartObject(miR = miR, mRNA = mRNA) -> MAE

e_list -> MAE@metadata$elist
dloadGMT(MAE, speciesInitial = "Mm") -> MAE

MAE@metadata$sigwiki <- EnrichWiki(method = "c",
e_list = MAE@metadata$elist,
orgDB = org.Mm.eg.db, 
path_gene = MAE@ExperimentList$path_gene, 
path_name = MAE@ExperimentList$path_name, 
ID = "ENTREZID", 
universe = MAE@ExperimentList$path_gene$gene)

SavePlots(largeList = MAE@metadata$sigwiki, maxInt = 5, 
quickType = Quickdot, fileType = "jpg") -> P
}
