% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantVals.R
\name{significantVals}
\alias{significantVals}
\title{significantVals}
\usage{
significantVals(MAE, method = '', geneList, maxVal, stringVal = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment to store the output of significantVals. It is
recommended to use the MAE used in the genesList.}

\item{method}{Either "c" or "s", respectively for combined or separated
analysis.}

\item{geneList}{A list of nested dataframes if "c" analysis is used or a list
of lists of nested dataframes if "s" is used. This will be the output from
them genesList function. The resulting list will be found as metadata, in
the MAE used in the genesList function.}

\item{maxVal}{Numeric value which represents the maximum cut off value
for significance e.g. 0.05.}

\item{stringVal}{Character. Common DE result type which is found in all
nested dataframes. This will be used for filtration e.g. pval, adjPval or
qval. Make sure the spelling matches the colnames for each sample.}
}
\value{
A list of dataframes with only significantly differentially
expressed genes. Output will be stored as metadata within the input MAE.
}
\description{
Filters out genes in each nested dataframe which are not deemed
significantly differentially expressed. Each sample will be filtered
independently.
}
\examples{
miR <- mm_miR

mRNA <- mm_mRNA

Data <- startObject(miR = miR, mRNA = mRNA)

Data <- combineGenes(MAE = Data, miR_data = assay(Data, 1),
                     mRNA_data = assay(Data, 2))

Data <- genesList(MAE = Data, method = 'c', genetic_data = assay(Data, 3),
                  timeString = 'D')

Data <- significantVals(MAE = Data, method = 'c',
                        geneList = metadata(Data)[[1]],
                        maxVal = 0.05, stringVal = "adjPVal")
}
