% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignificantVals.R
\name{SignificantVals}
\alias{SignificantVals}
\title{SignificantVals}
\usage{
SignificantVals(method = '', geneList, maxVal, stringVal = '')
}
\arguments{
\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{geneList}{A list of nested dataframes if 'c' or a list of lists
with nested dataframes if 's'.}

\item{maxVal}{Integer. Rows will be kept only if they are
lower than this value.}

\item{stringVal}{Character. Common resulttype in all nested
dataframes which should be point of filtration e.g. pval, adjPval,
qval. Make sure this matches the colnames.}
}
\value{
A list in a similair structure but with only significantly
differentially expressed genes.
}
\description{
Filter out genes in each nested dataframe which are not deemed
significantly differentially expressed.
}
\examples{
miR <- mm_miR
mRNA <- mm_mRNA
CombineGenes(miR_data = miR, mRNA_data = mRNA) -> genetic_data
GenesList(method = 'c', genetic_data = genetic_data,
timeString = 'D') -> genelist
SignificantVals(method = 'c', geneList = genelist, maxVal = 0.05,
stringVal = 'adjPVal') -> filtered_genelist
}
