% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantVals.R
\name{significantVals}
\alias{significantVals}
\title{significantVals}
\usage{
significantVals(MAE, method = '', geneList, maxVal, stringVal = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{method}{Respectively either 'c' or 's' for combined or separated
analysis.}

\item{geneList}{A list of nested dataframes if 'c' or a list of lists
with nested dataframes if 's'.}

\item{maxVal}{Integer. Rows will be kept only if they are
lower than this value.}

\item{stringVal}{Character. Common resulttype in all nested
dataframes which should be point of filtration e.g. pval, adjPval,
qval. Make sure this matches the colnames.}
}
\value{
A list in a similair structure but with only significantly
differentially expressed genes which will also be stores in the
metadata area of an MAE object.
}
\description{
Filter out genes in each nested dataframe which are not deemed
significantly differentially expressed.
}
\examples{
miR <- mm_miR
mRNA <- mm_mRNA
Data <- startObject(miR = miR, mRNA = mRNA)

Data <- combineGenes(MAE = Data, miR_data = assay(Data, 1),
                     mRNA_data = assay(Data, 2))

Data <- genesList(MAE = Data, method = 'c', genetic_data = assay(Data, 3),
                  timeString = 'D')

Data <- significantVals(MAE = Data, method = 'c',
                        geneList = metadata(Data)[[1]],
                        maxVal = 0.05, stringVal = "adjPVal")
}
