% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startObject.R
\name{startObject}
\alias{startObject}
\title{startObject}
\usage{
startObject(miR, mRNA)
}
\arguments{
\item{miR}{microRNA dataframe/ matrix. Rows should be miR gene names which
use the TimiRGeN friendly naming system. Columns should be results of
DE and time points.}

\item{mRNA}{mRNA dataframe/ matrix. Rows should be mRNA gene names.
Columns should be results of DE and time points.}
}
\value{
MultiAssayExperiment containing miR and mRNA data stored as assays.
Use assays(MAE, i) or MAE[[i]] to access assays. Use metadata(MAE)[[i]] to
access metadata. Use experiments(MAE)[[i]] to access experiments.
}
\description{
Creates a MultiAssayExperiment (MAE) from miR and mRNA
dataframes. MAE's will be the constant object used throughout TimiRGeN. The
input dataframes should contain rows as genes, and results from differential
expression (DE) as columns. Columns should also indicate the time point
related to each sample. Row names and column names must adhere to TimiRGeN
friendly nomenclature. Please do read the vignette for a full description of
the required nomenclature.
}
\examples{
data(mm_miR)

data(mm_mRNA)

Data <- startObject(miR = mm_miR, mRNA = mm_mRNA)
}
