% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnPercent.R
\name{turnPercent}
\alias{turnPercent}
\title{turnPercent}
\usage{
turnPercent(MAE, wikiMatrix, rowInt)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{wikiMatrix}{Matrix of wikipathways and samples.}

\item{rowInt}{Which row contains the total number of genes per wikipathway?}
}
\value{
A percentage matrix.
}
\description{
Converts integers into percentages, within a matrix.
}
\examples{
library(MultiAssayExperiment)
MAE <- MultiAssayExperiment()
metadata(MAE)["e_list"] <- e_list
metadata(MAE)["w_list"] <- w_list[1:10]
MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE,
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)
}
