% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TurnPercent.R
\name{TurnPercent}
\alias{TurnPercent}
\title{TurnPercent}
\usage{
TurnPercent(MAE, wikiMatrix, rowInt)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{wikiMatrix}{Matrix of wikipathways and samples.}

\item{rowInt}{Which row contains the total number of genes per wikipathway?}
}
\value{
A percentage matrix.
}
\description{
Converts integers into percentages, within a matrix.
}
\examples{
library(MultiAssayExperiment)
MAE <- MultiAssayExperiment()
metadata(MAE)["e_list"] <- e_list
metadata(MAE)["w_list"] <- w_list[1:10]
MAE <- WikiMatrix(MAE, ID_list = metadata(MAE)[[1]], 
                  wp_list = metadata(MAE)[[2]])
 
MAE <- TurnPercent(MAE = MAE, 
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)
}
