% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnPercent.R
\name{turnPercent}
\alias{turnPercent}
\title{turnPercent}
\usage{
turnPercent(MAE, wikiMatrix, rowInt)
}
\arguments{
\item{MAE}{MultiAssayExperiment object which will have the output of
turnPercent added as an assay. It is recommended to use the MAE produced by
the wikiMatrix function.}

\item{wikiMatrix}{Matrix of wikipathways and samples. This should be stored
as an assay in the MAE used in the wikiMatrix function.}

\item{rowInt}{Which row contains the total number of genes per wikipathway?
This will be 1+ the number of samples in your input data. For example our
combined miR-mRNA analysis in the example has rowInt = 6 because our example
has 5 time points.}
}
\value{
A percentage matrix.
}
\description{
Genes found in common between the input data and each
pathway are normalised by percentages to normalise for pathway size. Stores
output as an assay in a MAE object.
}
\examples{
MAE <- MultiAssayExperiment()

metadata(MAE)[["e_list"]] <- e_list_mouse

metadata(MAE)[["w_list"]] <- w_list_mouse[1:10]

MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])

MAE <- turnPercent(MAE = MAE,
                   wikiMatrix = assay(MAE, 1),
                   rowInt = 6)
}
