\name{w_list}
\alias{w_list}
\docType{data}
\title{
Wikipathways lists for mouse
}
\description{
List of entrezIDs associated with each wikipathway made for mouse. This dataset
will allow for functions to run faster during CMD check and vignette building.
}
\usage{data("w_list")}
\format{
The format is:
List of 226
$ WP1   : chr [1:20] "1071" "11303" "11806" "11808" ...
$ WP10  : chr [1:24] "11651" "12702" "14784" "16186" ...
$ WP103 : chr [1:15] "110196" "13121" "13360" "14137" ...
$ WP108 : chr [1:52] "107585" "107869" "109079" "109815" ...
$ WP113 : chr [1:52] "12159" "12387" "12393" "12399" ...
$ WP116 : chr [1:22] "12534" "12914" "13363" "13548" ...
$ WP1241: chr [1:17] "100727" "109785" "15275" "216558" ...
$ WP1242: chr [1:7] "108138" "14375" "17535" "19090" ...
$ WP1243: chr [1:21] "12675" "15519" "15977" "16150" ...
$ WP1244: chr [1:77] "100862456" "102216272" "102641086" "108143" ...
$ WP1245: chr [1:14] "12608" "13844" "14629" "14630" ...
$ WP1246: chr [1:32] "108071" "109550" "11459" "11514" ...
$ WP1247: chr [1:9] "108645" "11720" "12846" "140483" ...
$ WP1248: chr [1:59] "11946" "11947" "11950" "11951" ...
$ WP1249: chr [1:26] "110157" "11651" "12700" "12703" ...
$ WP1250: chr [1:3] "12858" "12868" "13850"
$ WP1251: chr [1:140] "102294" "104086" "107146" "110115" ...
$ WP1252: chr [1:2] "107146" "60525"
$ WP1253: chr [1:34] "100038882" "111654" "12265" "12702" ...
$ WP1254: chr [1:80] "11651" "11783" "11796" "11797" ...
$ WP1256: chr [1:9] "12675" "13844" "16150" "16151" ...
$ WP1257: chr [1:9] "16881" "17350" "17685" "17688" ...
$ WP1258: chr [1:13] "11920" "12190" "17535" "18971" ...
$ WP1259: chr [1:39] "11522" "11668" "12491" "12903" ...
$ WP1261: chr [1:46] "100042150" "11350" "11836" "11839" ...
$ WP1262: chr [1:5] "110198" "13077" "13849" "14862" ...
$ WP1263: chr [1:19] "12326" "12912" "14390" "15161" ...
$ WP1264: chr [1:13] "12846" "13076" "13077" "13078" ...
$ WP1265: chr [1:4] "11677" "16548" "20322" "230163"
$ WP1266: chr [1:61] "108015" "11438" "11444" "11516" ...
$ WP1269: chr [1:34] "110446" "11363" "11364" "11370" ...
$ WP1270: chr [1:62] "11504" "11595" "11647" "11651" ...
$ WP1271: chr [1:97] "106759" "110296" "11651" "11652" ...
$ WP1272: chr [1:31] "0" "103711" "10587" "114112" ...
$ WP1273: chr [1:27] "103711" "11689" "12359" "12411" ...
$ WP1274: chr [1:40] "102294" "104086" "109672" "109754" ...
$ WP1396: chr [1:267] "104443" "107221" "107831" "107934" ...
$ WP1397: chr [1:188] "100129" "107173" "107515" "110326" ...
$ WP1398: chr [1:19] "107173" "217143" "224792" "236798" ...
$ WP1496: chr [1:41] "11783" "12015" "12018" "12043" ...
$ WP1498: chr [1:30] "110648" "11668" "11924" "12577" ...
$ WP150 : chr [1:41] "106344" "12544" "12545" "12566" ...
$ WP1500: chr [1:6] "110648" "14179" "15376" "17701" ...
$ WP151 : chr [1:68] "107746" "110157" "110351" "11651" ...
$ WP1526: chr [1:7] "140781" "140781   " "14463" "14613" ...
$ WP1560: chr [1:83] "107589" "107951" "108058" "110157" ...
$ WP157 : chr [1:51] "100042025" "100043349" "103988" "106557" ...
$ WP1596: chr [1:14] "14319" "14325" "15216" "16175" ...
$ WP163 : chr [1:93] "100043348" "100043876" "100502825" "105244208" ...
$ WP164 : chr [1:19] "14381" "14598" "14629" "14630" ...
$ WP166 : chr [1:18] "11783" "12122" "12366" "12367" ...
$ WP175 : chr [1:7] "11423" "12647" "12660" "13026" ...
$ WP1763: chr [1:293] "102060" "103551" "104156" "104248" ...
$ WP1770: chr [1:52] "107869" "108037" "11720" "12012" ...
$ WP1771: chr [1:14] "12651" "12660" "13026" "18618" ...
$ WP18  : chr [1:9] "11655" "11656" "12892" "14151" ...
$ WP189 : chr [1:230] "11540" "11541" "11542" "11548" ...
$ WP190 : chr [1:88] "104394" "105988" "107995" "11350" ...
$ WP193 : chr [1:34] "107746" "109700" "109905" "110157" ...
$ WP1983: chr [1:42] "110351" "11603" "11735" "11804" ...
$ WP200 : chr [1:16] "109828" "110382" "12259" "12260" ...
$ WP201 : chr [1:11] "12387" "14165" "15205" "18096" ...
$ WP202 : chr [1:20] "107765" "11487" "11910" "13684" ...
$ WP2067: chr [1:44] "109575" "12154" "12156" "12159" ...
$ WP207 : chr [1:12] "11539" "11540" "11541" "11542" ...
$ WP2074: chr [1:101] "107771" "11852" "12005" "12006" ...
$ WP2075: chr [1:75] "11487" "11491" "11783" "11785" ...
$ WP2076: chr [1:13] "105980076" "14056" "15902" "17258" ...
$ WP2079: chr [1:4] "14682" "15558" "16452" "20848"
$ WP2081: chr [1:11] "105980076" "108058" "12444" "13712" ...
$ WP2082: chr [1:12] "105980076" "15184" "17258" "17259" ...
$ WP2083: chr [1:10] "110834" "11438" "11444" "12568" ...
$ WP2084: chr [1:11] "15357" "16835" "19013" "19017" ...
$ WP2085: chr [1:10] "11350" "12443" "12447" "12530" ...
$ WP2140: chr [1:13] "11838" "12918" "12921" "14281" ...
$ WP2141: chr [1:18] "11549" "11838" "12323" "12918" ...
$ WP2152: chr [1:14] "12028" "12043" "12064" "12540" ...
$ WP216 : chr [1:45] "109272" "11459" "11464" "11465" ...
$ WP2170: chr [1:6] "13609" "13610" "20397" "20698" ...
$ WP2185: chr [1:175] "100043714" "100046796" "100862375" "102216272" ...
$ WP222 : chr [1:29] "12504" "12977" "12981" "12985" ...
$ WP2271: chr [1:10] "12475" "12514" "12522" "12524" ...
$ WP2292: chr [1:193] "100039053" "100039863" "100039939" "100040048" ...
$ WP2309: chr [1:840] "100048834" "100862175" "101055731" "101055745" ...
$ WP2310: chr [1:315] "101543" "103967" "104027" "105782" ...
$ WP2316: chr [1:82] "100040843" "103968" "104086" "108078" ...
$ WP2318: chr [1:10] "11364" "11370" "11409" "12894" ...
$ WP232 : chr [1:92] "100043507" "100986" "104110" "104111" ...
$ WP234 : chr [1:70] "11607" "11608" "11609" "12000" ...
$ WP2375: chr [1:7] "18999" "20674" "53607" "53817" ...
$ WP238 : chr [1:11] "12097" "12842" "12907" "18383" ...
$ WP240 : chr [1:15] "107239" "109900" "11484" "11564" ...
$ WP2432: chr [1:100] "11595" "11829" "11846" "11848" ...
$ WP246 : chr [1:188] "100039786" "100043702" "100043714" "103583" ...
$ WP251 : chr [1:28] "109880" "110157" "11836" "13712" ...
$ WP252 : chr [1:112] "100043349" "104156" "107951" "108767" ...
$ WP2561: chr ""
$ WP2573: chr [1:72] "104027" "110829" "11603" "11651" ...
$ WP258 : chr [1:151] "100043674" "103573" "104394" "108058" ...
[list output truncated]
}
\details{
A Large list of each wikipathway and associated genes.
}
\source{
https://www.wikipathways.org/index.php/Download_Pathways
}
\references{
Denise N Slenter et al. “WikiPathways: a multifaceted pathway database b
ridging metabolomics to other omics research”.
In: Nucleic acids research 46.D1 (2017), pp. D661–D667.
}
\examples{
data(w_list) -> w_list
}
\keyword{datasets}
