\name{w_list_human}
\alias{w_list_human}
\docType{data}
\title{
w_list_human
}
\description{
List of human wikipathways.
}
\usage{data("w_list_human")}
\format{
The format is:
List of 1158
$ WP100 : chr [1:22] "102724197" "221357" "2539" "2678" ...
$ WP106 : chr [1:12] "1615" "18" "189" "2571" ...
$ WP107 : chr [1:52] "101930123" "10209" "10605" "107984923" ...
$ WP111 : chr [1:103] "10063" "10476" "10632" "10975" ...
$ WP117 : chr [1:103] "10888" "11245" "1129" "1131" ...
$ WP12  : chr [1:16] "1513" "3454" "3456" "3690" ...
$ WP127 : chr [1:40] "1439" "161742" "2002" "207" ...
$ WP129 : chr [1:30] "10893" "118856" "4312" "4313" ...
$ WP134 : chr [1:7] "22934" "2539" "25796" "5226" ...
$ WP136 : chr [1:8] "1066" "2098" "221223" "3988" ...
$ WP138 : chr [1:90] "1026" "10273" "10399" "10401" ...
$ WP1403: chr [1:70] "101930123" "1026" "10645" "126129" ...
$ WP1422: chr [1:30] "10087" "10558" "125981" "130367" ...
$ WP1423: chr [1:13] "127550" "2523" "2583" "29906" ...
$ WP1424: chr [1:21] "10317" "10610" "10690" "2523" ...
$ WP1425: chr [1:12] "10140" "10766" "4086" "4089" ...
$ WP143 : chr [1:34] "1120" "1374" "1375" "1376" ...
$ WP1433: chr [1:41] "10392" "10910" "10980" "114548" ...
$ WP1434: chr [1:13] "1147" "3551" "3685" "3690" ...
$ WP1438: chr [1:13] "3802042" "407042" "596" "956527" ...
$ WP1449: chr [1:145] "10000" "10193" "10333" "10454" ...
$ WP1455: chr [1:11] "10066" "121278" "3553" "3554" ...
$ WP1471: chr [1:36] "10325" "10670" "115265" "1978" ...
$ WP1495: chr [1:3] "4524" "6470" "6472"
$ WP15  : chr [1:90] "100129518" "102724197" "102724560" "10549" ...
$ WP1528: chr [1:24] "1432" "1489" "183" "1906" ...
$ WP1530: chr [1:98] "100616357" "10111" "1017" "1019" ...
$ WP1531: chr [1:10] "120227" "1591" "1593" "1594" ...
$ WP1533: chr [1:54] "100129518" "102724560" "107986404" "12" ...
$ WP1539: chr [1:24] "1387" "1432" "207" "2247" ...
$ WP1541: chr [1:47] "100271849" "10514" "10891" "133522" ...
$ WP1544: chr [1:104] "10014" "1022" "1025" "10454" ...
$ WP1545: chr [1:68] "100126302" "100126317" "100313923" "100422842" ...
$ WP1559: chr [1:15] "207" "208" "406904" "406905" ...
$ WP1584: chr [1:23] "122809" "23533" "2475" "2710" ...
$ WP1589: chr [1:10] "102724560" "1051" "10840" "128" ...
$ WP1591: chr [1:47] "1482" "1499" "150572" "2065" ...
$ WP1600: chr [1:6] "1548" "1555" "2328" "316" ...
$ WP1601: chr [1:34] "10057" "10257" "1066" "10864" ...
$ WP1602: chr [1:21] "1020" "108" "1136" "1137" ...
$ WP1603: chr [1:4] "1136" "1143" "775" "8913"
$ WP1604: chr [1:17] "10599" "107987478" "107987479" "1244" ...
$ WP167 : chr [1:25] "100137049" "102724197" "1800" "239" ...
$ WP170 : chr [1:38] "10062" "190" "2063" "2099" ...
$ WP1742: chr [1:19] "1026" "1029" "1647" "25" ...
$ WP176 : chr [1:73] "100129518" "102724560" "10588" "10797" ...
$ WP1772: chr [1:93] "10017" "10018" "1029" "1616" ...
$ WP1775: chr [1:288] "10111" "1017" "10197" "10213" ...
$ WP1780: chr [1:104] "10057" "10058" "10060" "10197" ...
$ WP1781: chr [1:13] "1650" "177" "3146" "351" ...
$ WP1784: chr [1:51] "100506658" "10059" "10134" "118" ...
$ WP1785: chr [1:125] "10195" "10206" "10299" "1081" ...
$ WP1788: chr [1:42] "10005" "10499" "10599" "10858" ...
$ WP1789: chr [1:3] "10642" "10643" "10644"
$ WP179 : chr [1:120] "1017" "1019" "1021" "1022" ...
$ WP1793: chr [1:86] "1000" "1001" "1002" "1003" ...
$ WP1794: chr [1:128] "10461" "1048" "105369247" "10544" ...
$ WP1795: chr [1:24] "10654" "10682" "1595" "1717" ...
$ WP1797: chr [1:57] "10498" "10499" "10743" "10891" ...
$ WP1798: chr [1:55] "10584" "10877" "10878" "110384692" ...
$ WP1799: chr [1:62] "10000" "117145" "1326" "1445" ...
$ WP1800: chr [1:12] "10369" "11094" "27092" "55799" ...
$ WP1802: chr [1:13] "302" "3273" "5054" "5055" ...
$ WP1803: chr [1:49] "10459" "10714" "11201" "124739" ...
$ WP1804: chr [1:8] "10973" "121642" "221120" "4255" ...
$ WP1809: chr [1:27] "10125" "10235" "11343" "139189" ...
$ WP1811: chr [1:95] "101929876" "11224" "116832" "140032" ...
$ WP1812: chr [1:122] "101929876" "10480" "107984923" "11224" ...
$ WP1813: chr [1:95] "101929876" "11224" "116832" "140032" ...
$ WP1815: chr [1:106] "10019" "1017" "1020" "10362" ...
$ WP1818: chr [1:38] "10544" "2147" "2149" "2152" ...
$ WP1820: chr [1:29] "10052" "10804" "1173" "1211" ...
$ WP1822: chr [1:398] "10001" "100129543" "100129842" "100131980" ...
$ WP1823: chr [1:9] "2316" "2811" "2812" "2814" ...
$ WP1826: chr [1:33] "10451" "10630" "146850" "2207" ...
$ WP1829: chr [1:107] "100528032" "10219" "10225" "102725035" ...
$ WP183 : chr [1:64] "10197" "3014" "3015" "3105" ...
$ WP1831: chr [1:109] "10411" "10554" "10681" "107" ...
$ WP1833: chr [1:65] "100505984" "1282" "1284" "1285" ...
$ WP1835: chr [1:53] "103" "10379" "10410" "10581" ...
$ WP1836: chr [1:86] "10107" "10346" "10379" "10437" ...
$ WP1838: chr [1:7] "3552" "3553" "3606" "4790" ...
$ WP1839: chr [1:127] "10068" "10197" "10213" "10392" ...
$ WP1840: chr [1:43] "1398" "1399" "1437" "1438" ...
$ WP1841: chr [1:44] "10000" "10018" "10971" "140735" ...
$ WP1842: chr [1:40] "10112" "1062" "10749" "11004" ...
$ WP1843: chr [1:96] "10048" "101928143" "10752" "11280" ...
$ WP1845: chr [1:31] "1385" "1386" "1432" "1845" ...
$ WP185 : chr [1:103] "10000" "100505984" "10298" "10451" ...
$ WP1850: chr [1:15] "10434" "114757" "115677" "1785" ...
$ WP1852: chr [1:17] "1352" "1355" "1371" "210" ...
$ WP1857: chr [1:75] "10243" "105372824" "10560" "107986404" ...
$ WP1858: chr [1:148] "10000" "1017" "1019" "10197" ...
$ WP1859: chr [1:181] "10121" "10133" "10142" "1017" ...
$ WP186 : chr [1:13] "10111" "10714" "25788" "4361" ...
$ WP1861: chr [1:28] "1022" "2068" "2071" "22916" ...
$ WP1862: chr [1:11] "100913187" "103" "104" "10930" ...
$ WP1865: chr [1:28] "1000" "1002" "100271849" "1013" ...
$ WP1866: chr [1:37] "1291" "1292" "1297" "1298" ...
[list output truncated]
}
\details{
A Large list of each wikipathway and associated genes.
}
\source{
https://www.wikipathways.org/index.php/Download_Pathways
}
\references{
Denise N Slenter et al. “WikiPathways: a multifaceted pathway database b
ridging metabolomics to other omics research”.
In: Nucleic acids research 46.D1 (2017), pp. D661–D667.
}
\examples{
data(w_list_human) -> w_list_human
}
\keyword{datasets}

