% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiList.R
\name{wikiList}
\alias{wikiList}
\title{wikiList}
\usage{
wikiList(MAE, stringSpecies = '', stringSymbol = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the results from
wikiList. It is recommended to use the same MAE which stores output from the
dloadGmt function.}

\item{stringSpecies}{Full species name to decide which wikipathways to
download. 'Homo sapiens' to download human pathways or 'Mus musculus'
to download mouse pathways.}

\item{stringSymbol}{Type of gene ID to retrieve e.g. 'En' for ensemble gene
IDs or 'L' for entrezgene IDs.}
}
\value{
List of wikipathways and associated genes saved as as strings. Output
will be stored as metadata in the input MAE.
}
\description{
Provides a list of wikipathways specific for a species, and each
gene found to be within each pathway. wikiList will download and process a
large amount of data from the wikipathways website so this may take some
time to complete.
}
