% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikiList.R
\name{WikiList}
\alias{WikiList}
\title{WikiList}
\usage{
WikiList(stringSpecies = '', stringSymbol = '')
}
\arguments{
\item{stringSpecies}{species wikipathways relate to e.g. 'Homo sapiens'}

\item{stringSymbol}{Type of gene code e.g. 'H'}
}
\value{
List of wikipathways and associated genes as a list of strings.
}
\description{
Provides a list of wikipathways and each gene found to be
within that wikipathway as the value.
}
\details{
install rWikiPathways.
}
