% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiList.R
\name{wikiList}
\alias{wikiList}
\title{wikiList}
\usage{
wikiList(MAE, stringSpecies = '', stringSymbol = '')
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{stringSpecies}{species wikipathways relate to e.g. 'Homo sapiens'}

\item{stringSymbol}{Type of gene code e.g. 'H', 'En', 'L'.}
}
\value{
List of wikipathways and associated genes as a list of strings.
}
\description{
Provides a list of wikipathways and each gene found to be
within that wikipathway as the value.
}
\examples{
\donttest{
MAE <- MultiAssayExperiment()

MAE <- wikiList(MAE, stringSpecies = 'Homo sapiens', stringSymbol = 'En')
}
}
