% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiMatrix.R
\name{wikiMatrix}
\alias{wikiMatrix}
\title{wikiMatrix}
\usage{
wikiMatrix(MAE, ID_list, wp_list)
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{ID_list}{list of gene names as strings.}

\item{wp_list}{list of list of wikipathways with gene names as strings.}
}
\value{
A matrix of timepoints and wikipathways.
}
\description{
Creates a matrix which shows how many genes from the data are
found in each wikipathway for a specific species.
}
\examples{
library(MultiAssayExperiment)
MAE <- MultiAssayExperiment()
metadata(MAE)[["ID_list"]] <- e_list
metadata(MAE)[["w_list"]] <- w_list[1:10]
MAE <- wikiMatrix(MAE, ID_list = metadata(MAE)[[1]],
                  wp_list = metadata(MAE)[[2]])
}
