% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikimRNA.R
\name{WikimRNA}
\alias{WikimRNA}
\title{WikimRNA}
\usage{
WikimRNA(mRNA_express, SingleWiki)
}
\arguments{
\item{mRNA_express}{Output from the Express function on mRNA data.}

\item{SingleWiki}{Output from ReduceWiki function on a Wikipathway
of interest.}
}
\value{
A dataframe of mRNA entries which are also found in the wikipathway
of interest.
}
\description{
Identify which mRNAs are in common with the genes found in the
wikipathway of interest.
}
\examples{
library(biomaRt)
mm_mRNA -> mRNA
getIDs_mRNA_mouse(mRNA)
Express(df = mRNA, dataType = 'Log2FC', genes_ID = mRNA_entrez,
idColumn = 'GENENAME') -> mRNA_express
path_data <- data.frame("wpid" = c(rep("WP571", 6)),
"gene" = c(16175, 12370,26419, 19249, 19645, 18479),
"name" = c(rep("Fas pathway and Stress induction of HSP regulation", 6)))
ReduceWiki(path_data = path_data,
stringWiki = 'Fas pathway and Stress induction of
HSP regulation') -> singlewiki
WikimRNA(mRNA_express = mRNA_express,
SingleWiki = singlewiki) -> GenesofInteres
}
