% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikimRNA.R
\name{WikimRNA}
\alias{WikimRNA}
\title{WikimRNA}
\usage{
WikimRNA(mRNA_express, SingleWiki)
}
\arguments{
\item{mRNA_express}{Output from the Express function on mRNA data.}

\item{SingleWiki}{Output from ReduceWiki function on a Wikipathway
of interest.}
}
\value{
A dataframe of mRNA entries which are also found in the wikipathway
of interest.
}
\description{
Identify which mRNAs are in common with the genes found in the
wikipathway of interest.
}
\examples{
library(biomaRt)
mm_mRNA -> mRNA
StartObject(miR = NULL, mRNA = mRNA) -> MAE
getIDs_mRNA_mouse(MAE, MAE@ExperimentList$mRNA) -> MAE
Express(df = MAE@ExperimentList$mRNA, dataType = 'Log2FC', 
genes_ID = MAE@ExperimentList$mRNA_entrez,
idColumn = 'GENENAME') -> mRNA_express
dloadGMT(MAE = MAE, speciesInitials = "Mm") -> MAE
ReduceWiki(path_data = MAE@ExperimentList$path_data,
stringWiki = 'TGF Beta Signaling Pathway') -> singlewiki
Express(df = MAE@ExperimentList$mRNA, dataType = 'Log2FC',
genes_ID = MAE@ExperimentList$mRNA_entrez,
idColumn = 'GENENAME') -> MAE@ExperimentList$mRNA_express
WikimRNA(mRNA_express = MAE@ExperimentList$mRNA_express,
SingleWiki = singlewiki) -> MAE@ExperimentList$GenesofInterest
}
