% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiMrna.R
\name{wikiMrna}
\alias{wikiMrna}
\title{wikiMrna}
\usage{
wikiMrna(MAE, mRNA_express, singleWiki, stringWiki='')
}
\arguments{
\item{MAE}{MultiAssayExperiment object.}

\item{mRNA_express}{Output from the Express function on mRNA data.}

\item{singleWiki}{Output from ReduceWiki function on a Wikipathway
of interest.}

\item{stringWiki}{Name of the pathway of interest.}
}
\value{
A dataframe of mRNA entries which are also found in the wikipathway
of interest.
}
\description{
Identify which mRNAs are in common with the genes found in the
wikipathway of interest.
}
\examples{
library(biomaRt)
library(MultiAssayExperiment)
miR <- mm_miR[1:100,]
mRNA <- mm_mRNA[1:200,]
MAE <- startObject(miR = miR, mRNA = mRNA)
MAE <- getIdsMirMouse(MAE, assay(MAE, 1))
MAE <- getIdsMrnaMouse(MAE, assay(MAE, 2), "useast")
MAE <- dloadGmt(MAE = MAE, speciesInitials = "Mm")

MAE <- reduceWiki(MAE, path_data = assay(MAE, 11),
                  stringWiki = 'TGF Beta Signaling Pathway')

MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
               genes_ID = assay(MAE, 7),
               idColumn = 'GENENAME',
               name = "mRNA_log")

MAE <- wikiMrna(MAE, mRNA_express = assay(MAE, 13),
                singleWiki = assay(MAE, 12),
                stringWiki = 'TGF Beta Signaling Pathway')
}
