% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiMrna.R
\name{wikiMrna}
\alias{wikiMrna}
\title{wikiMrna}
\usage{
wikiMrna(MAE, mRNA_express, singleWiki, stringWiki='')
}
\arguments{
\item{MAE}{MultiAssayExperiment which will store the results of wikiMrna.
It is recommended to use the same MAE which stores output from the
diffExpressRes and reduceWiki functions.}

\item{mRNA_express}{Dataframe from the diffExpressRes function used on the
input mRNA data. This should be found as an assay within the MAE used in the
diffExpressRes function.}

\item{singleWiki}{Dataframe containing information about only one pathway.
This is output from the reduceWiki function. This should be found as an
assay within the MAE used in the reduceWiki function.}

\item{stringWiki}{Name of the pathway of interest. Should be the same
as the stringWiki parameter from the reduceWiki function.}
}
\value{
A dataframe which only contains mRNAs which are found in both the
input data and the wikipathway of interest. Output will be stored as an assay
in the input MAE.
}
\description{
Identify genes that are in common in both the wikipathway of
interest and the significantly differentially expressed input mRNAs.
}
\examples{
library(org.Mm.eg.db)

miR <- mm_miR[1:200,]

mRNA <- mm_mRNA[401:600,]

MAE <- startObject(miR = miR, mRNA = mRNA)

MAE <- getIdsMirMouse(MAE, assay(MAE, 1))

MAE <- getIdsMrnaMouse(MAE, assay(MAE, 2), "useast")

MAE <- dloadGmt(MAE = MAE, speciesInitials = "Mm")

MAE <- reduceWiki(MAE, path_data = assay(MAE, 11),
                  stringWiki = 'TGF Beta Signaling Pathway')

MAE <- diffExpressRes(MAE, df = assay(MAE, 2), dataType = 'Log2FC',
               genes_ID = assay(MAE, 7),
               idColumn = 'GENENAME',
               name = "mRNA_log")

MAE <- wikiMrna(MAE, mRNA_express = assay(MAE, 13),
                singleWiki = assay(MAE, 12),
                stringWiki = 'TGF Beta Signaling Pathway')
}
